==> 0_helpers.R <==
#' # Helper functions used throughout {.tabset .tabset-sticky}
#' Documentation on the functions is interspersed through code comments
#'
#' ## Set some options
#' do net set strings as factors
options(stringsAsFactors = FALSE)
#' show four significant digits
options(digits = 4)
#' tend not to show scientific notation
options(scipen = 7)
#' make output a bit wider
options(width = 110)
#' set a seed to make analyses depending on random number generation reproducible
set.seed(1710)


#' ## Load packages
#' generate the site
library(rmarkdown)
#' set options for chunks
library(knitr)
#' formr utility package to generate e.g. the bibliography
library(formr)
#' pretty-printed output
library(pander)
#' tidyverse date times
library(lubridate)
#' tidyverse strings
library(stringr)
#' extractor functions for models
library(broom)
library(broom.mixed)
#' marginal effects plots for regressions
library(effects)
#' grammar of graphics plots
library(ggplot2)
#' enhanced version of data.frame
library(data.table)
#' linear mixed effects models
library(lme4)
#' interpret linear mixed effects models output
library(lmerTest)
#' effect displays for all sort of models
library(effects)
#' simple, consistent wrappers for common string operations
library(stringr)
#' tools for working with categorical variables
library(forcats)
#' provides a funtion to format R source corde
library(formatR)
#' loads specific r tools
library(devtools)
#' imports IFLS data
library(haven)
#' package to perform factor analysis
library(GPArotation)
#' another package to perform factor analysis
library(psych)
#' yet another package to perform factor analysis
library(lavaan)
#' many usefull functions for data analysis
library(Hmisc)
#' all colors are from here
library(RColorBrewer)
#' tidyverse: transform data wide to long
library(tidyr)
#' tidyverse-style data wrangling. has a lot of naming conflicts, so always load last
library(dplyr)
#' make comparisons between different birth orders
library(coefplot)
#' nice overview of raw data
library(codebook)
#' used to compute effect sizes for t tests
library(effsize)

#' ## Spin R files
#' R scripts can be documented in markdown using Roxygen comments, as demonstrated here
#' This function turns all R files (that don't have an Rmd file of the same name and that don't start with an underscore _) into HTML pages
spin_R_files_to_site_html = function() {
  library(knitr)
  all_Rs = c(list.files(pattern = "^[^_].+\\.R$"), ".Rprofile")
  component_Rmds = list.files(pattern = "^_.+\\.Rmd$")
  temporary_Rmds = c()
  for (i in seq_along(all_Rs)) {
    if(all_Rs[i] == ".Rprofile") {
      Rmd_file = ".Rprofile.Rmd"
    } else {
      Rmd_file = paste0(all_Rs[i], "md")
    }
    if (!file.exists(Rmd_file)) {
      next_document = length(temporary_Rmds) + 1
      temporary_Rmds[next_document] = spin(all_Rs[i], knit = FALSE, envir = new.env(), format = "Rmd")
      prepended_yaml = paste0(c("---
output:
  html_document:
    code_folding: 'show'
---

", readLines(temporary_Rmds[next_document])), collapse = "\n")
      cat(prepended_yaml, file = temporary_Rmds[next_document])
    }
  }
  components_and_scripts = c(temporary_Rmds, component_Rmds)
  for (i in seq_along(components_and_scripts)) {
    opts_chunk$set(eval = FALSE, cache = FALSE)
    # if we call render_site on the .R file directly it adds a header I don't like
    rmarkdown::render_site(components_and_scripts[i], quiet = TRUE)
  }
  opts_chunk$set(eval = TRUE, cache = TRUE)
  unlink(temporary_Rmds)
}

#' ## Output options
#' use pander to pretty-print objects (if possible)
pander_handler = function(x, ...) {
  anyS3method = function(x) {
    classes = class(x)
    any(sapply(classes, FUN = function(classes) { !is.null(getS3method('pander',classes,TRUE)) } ))
  }
  if ("knit_asis" %in% class(x)) {
    x
  } else if (is.data.table(x)) {
    ""
    # don't ever print stupid data tables
  } else if (anyS3method(x)) {
    pander(x, row.names = F, ...)
  } else if (isS4(x)) {
    show(x)
  } else {
    print(x)
  }
}

opts_chunk$set(
  render = pander_handler
)

#' do not show warnings
opts_chunk$set(warning = FALSE)

#' don't split tables, scroll horizontally
panderOptions("table.split.table", Inf)

#' ## Functions used

#' Function to extract month from dataset
pad_month = function(x) { str_pad(x, width = 2, side = "left", pad = "0")}

#' Function to calculate the birthdate out of all available informations for one individual
all_available_info_birth_date = function(byear, bmonth, bday = NULL) {
  if(!is.null(bday)) {
    bday = paste0("-", pad_month(bday))
  } else {
    bday = ""
  }
  ifelse(is.na(byear), NA,
         paste0(byear, "-", pad_month(bmonth), pad_month(bday)))
  # can yield 2016-NA-NA
  #           2016-01-NA
  #           2016-01-01
  #           2016-01
}

#' Function to calculate the birthorder based on the siblings still alive at the time of birth
older_sibs_alive_and_dependent = function(byear, dyear) {
  sibs = length(byear)
  older_sibs_alive_and_dependent = integer(length=sibs) + 1
  for(i in 1:sibs) {
    older_sibs = byear <= byear[i] # not using < because of twins
    older_sibs[i] = F # minus self
    my_sibs = sum(older_sibs,na.rm = T) # minus self
    if(my_sibs > 0) {
      sib_births = byear[ which(older_sibs) ]
      sib_deaths = dyear[ which(older_sibs) ]
      my_sibs = my_sibs -
        sum(
          # sib_births < (byear[i] - 5) | # others born more than 5y earlier than me  # 10 seconds of 17
          (sib_deaths <= byear[i]) # died before my birth
          ,na.rm=T)
      older_sibs_alive_and_dependent[i] = my_sibs
    }
  }
  older_sibs_alive_and_dependent
}

#' set apa-conform theme for all graphs
apatheme=theme_bw()+
  theme(panel.grid.major=element_blank(),
        panel.grid.minor=element_blank(),
        panel.border=element_blank(),
        axis.line=element_line(),
        text = element_text(size=20))

#' function to plot gender plots
plot_gender = function(data = birthorder){
  plot = ggplot(data, aes(male, outcome)) +
    geom_jitter(width = 0.3) +
    stat_summary(fun.data = mean_cl_normal, geom = "errorbar", color = "red", width = 0.3)
  plot
}

#' function to plot age plots
plot_age = function(data = birthorder){
  plot = ggplot(data, aes(age, outcome)) +
    geom_jitter() +
    geom_smooth(stat = "smooth")
  plot
}

#' function used to do all birth order effect plots except for linear birth order effects
plot_birthorder = function(model, title = "", bo_var = "birth_order", separate = TRUE) {
  if(inherits(model, "merMod")) {
    varnames = names(model@frame)
  } else {
    varnames = names(model$model)
  }
  outcome = varnames[1]
  library(effects)
  library(tidyr)
  emm = allEffects(model, confidence.level = 0.995)
  bo_var = names(emm)[names(emm) %contains% bo_var]
  cemm = as.data.frame(emm[[bo_var]])
  if (separate != TRUE) {
    cemm = cemm %>% rename_("Birth order" = bo_var) %>% mutate(Sibship = "across")
  } else {
    cemm = cemm %>%
      separate_(bo_var, into = c("Birth order", "Sibship"), sep = "/")
    number = spread(as.data.frame(table(model.frame(model)[`bo_var`])), Var1, Freq)
    n2 = paste0("2 (", sum(number$`1/2`, number$`2/2`), ")", seperate="")
    n3 = paste0("3 (", sum(number$`1/3`, number$`2/3`, number$`3/3`), ")",
                seperate="")
    n4 = paste0("4 (", sum(number$`1/4`, number$`2/4`, number$`3/4`,
                           number$`4/4`), ")", seperate="")
    n5 = paste0("5 (", sum(number$`1/5`, number$`2/5`, number$`3/5`,
                           number$`4/5`, number$`5/5`), ")", seperate="")
    n5more = paste0(">5 (", sum(number$`1/>5`, number$`2/>5`, number$`3/>5`,
                                number$`4/>5`, number$`5/>5`, number$`>5/>5`), ")",
                    seperate="")
    cemm = cemm %>%
      mutate(Sibship = recode_factor(Sibship, "2" = `n2`, "3" = `n3`, "4" = `n4`,
                                     "5" = `n5`, ">5" = `n5more`))
  }
  plotx = ggplot(cemm, aes(factor(`Birth order`,levels = c("1", "2", "3", "4", "5", ">5")),
                           y = fit, ymax = upper, ymin = lower,
                           colour = `Sibship`, group = `Sibship`)) +
    geom_pointrange(stat = "identity", position = position_dodge(width = 0.5), size = 1) +
    geom_line(position = position_dodge(width = 0.5), size = 1) +
    labs(title = title) +
    xlab("Birth order") +
    apatheme +
    theme(text = element_text(size=20), axis.text.x = element_text(size = 15),
          axis.text.y = element_text(size = 15), legend.text = element_text(size = 15),
          legend.title = element_text(size = 15),
          legend.background = element_rect(fill="gray90", size=.5, linetype="dotted"),
          plot.title = element_text(hjust = 0)) +
    guides(colour=guide_legend(title = "Sibship")) +
    scale_fill_brewer(palette = "Set2") +
    scale_colour_brewer(palette = "Set2")
  print(plotx)
  assign(paste0("plot_", outcome, seperate=""),plotx,.GlobalEnv)
}

plot_birthorder2 = function(model, title = "", bo_var = "birth_order", separate = TRUE) {
  if(inherits(model, "merMod")) {
    varnames = names(model@frame)
  } else {
    varnames = names(model$model)
  }
  outcome = varnames[1]
  library(effects)
  library(tidyr)
  emm = allEffects(model, confidence.level = 0.995)
  bo_var = names(emm)[names(emm) %contains% bo_var]
  cemm = as.data.frame(emm[[bo_var]])
  if (separate != TRUE) {
    cemm = cemm %>% rename_("Birth order" = bo_var) %>% mutate(Sibship = "across")
  } else {
    cemm = cemm %>%
      separate_(bo_var, into = c("Birth order", "Sibship"), sep = "/")
    number = spread(as.data.frame(table(model.frame(model)[`bo_var`])), Var1, Freq)
    n2 = paste0("2 (", sum(number$`1/2`, number$`2/2`), ")", seperate="")
    n3 = paste0("3 (", sum(number$`1/3`, number$`2/3`, number$`3/3`), ")",
                seperate="")
    n4 = paste0("4 (", sum(number$`1/4`, number$`2/4`, number$`3/4`,
                           number$`4/4`), ")", seperate="")
    n5 = paste0("5 (", sum(number$`1/5`, number$`2/5`, number$`3/5`,
                           number$`4/5`, number$`5/5`), ")", seperate="")
    n5more = paste0(">5 (", sum(number$`1/>5`, number$`2/>5`, number$`3/>5`,
                                number$`4/>5`, number$`5/>5`, number$`>5/>5`), ")",
                    seperate="")
    cemm = cemm %>%
      mutate(Sibship = recode_factor(Sibship, "2" = `n2`, "3" = `n3`, "4" = `n4`,
                                     "5" = `n5`, ">5" = `n5more`))
  }
  plotx = ggplot(cemm, aes(`Birth order`, y = fit, ymax = upper, ymin = lower,
                           colour = `Sibship`, group = `Sibship`)) +
    geom_smooth(stat = "identity", position = position_dodge(width = 0.5), size = 1) +
    geom_line(position = position_dodge(width = 0.5), size = 1) +
    labs(title = title) +
    xlab("Birth order") +
    apatheme +
    theme(text = element_text(size=20), axis.text.x = element_text(size = 15),
          axis.text.y = element_text(size = 15), legend.text = element_text(size = 15),
          legend.title = element_text(size = 15),
          legend.background = element_rect(fill="gray90", size=.5, linetype="dotted"),
          plot.title = element_text(hjust = 0)) +
    guides(colour=guide_legend(title = "Sibship")) +
    scale_fill_brewer(palette = "Set2") +
    scale_colour_brewer(palette = "Set2")
  print(plotx)
  assign(paste0("plot_", outcome, seperate=""),plotx,.GlobalEnv)
}

#' function used to compare models
compare_models_markdown = function(m2_birthorder_linear) {
  formr::asis_knit_child('_test_outcome.Rmd')
  }

#' function used to run different birth order specifications
compare_birthorder_specs = function(m2_birthorder_linear) {
  formr::asis_knit_child('_compare_birthorder_specs.Rmd')
}


#' function used for imputation
mutate.mitml.list <- function(.data, ...) {
  r <- lapply(.data, FUN = mutate, ...)
  class(r) = class(.data)
  r
}

filter.mitml.list <- function(.data, ...) {
  r <- lapply(.data, FUN = filter, ...)
  class(r) = class(.data)
  r
}

tbl_df.mitml.list <- function(.data, ...) {
  r <- lapply(.data, FUN = tbl_df, ...)
  class(r) = class(.data)
  r
}

n_imputations.mitml = function(imps) {
  imps$iter$m
}
n_imputations.mids = function(imps) {
  imps$m
}
n_imputations.amelia = function(imps) {
  length(imps$imputations)
}
n_imputations = function(imps) { UseMethod("n_imputations") }

complete.mitml = mitml::mitmlComplete
complete.mids = mice::complete
complete = function(...) { UseMethod("complete") }

select.mitml.list <- function(.data, ...) {
  r <- lapply(.data, FUN = select, ...)
  class(r) = class(.data)
  r
}

group_by.mitml.list <- function(.data, ...) {
  r <- lapply(.data, FUN = group_by, ...)
  class(r) = class(.data)
  r
}

ungroup.mitml.list <- function(.data, ...) {
  r <- lapply(.data, FUN = ungroup, ...)
  class(r) = class(.data)
  r
}

summarise.mitml.list <- function(.data, ...) {
  r <- lapply(.data, FUN = summarise, ...)
  class(r) = class(.data)
  r
}

#' function used to compare models based on imputed data
compare_birthorder_imputed = function(data_used) {
  formr::asis_knit_child('_compare_birthorder_imputed.Rmd')
}
