==> 2_Inclusion.Rmd <==
---
title: "2_Inclusion"
author: "Laura Botzet & Ruben Arslan"
output: html_document
editor_options: 
  chunk_output_type: console
---
# <span style="color:#8DA0CB">Data Sampling</span> {.tabset}

## Helper
```{r helper}
source("0_helpers.R")
```

## Load data
```{r Load data}
### Import alldata used for birthorder
alldata_birthorder = readRDS("data/alldata_birthorder.rds")
```


## Birthorder {.active}

### Total
Number of individuals we were able to reconstruct the three different birth orders based on pregnancy and marriage history for 
```{r}
## Count Birthorder
n_birthorder = alldata_birthorder %>%
  summarise(Maternal_Birth_Order = sum(!is.na(birthorder_uterus_alive)),
            Maternal_Pregnancy_Order = sum(!is.na(birthorder_uterus_preg)),
            Parental_Full_Sibling_Birth_Order = sum(!is.na(birthorder_genes)),
            Naive_Maternal_Birth_Order = sum(!is.na(birthorder_naive)))
n_birthorder
```


### No multiple births
``` {r}
n_birthorder = alldata_birthorder %>%
  summarise(Maternal_Birth_Order = sum(!is.na(birthorder_uterus_alive) & any_multiple_birth != 1,
                                       na.rm = T),
            Maternal_Pregnancy_Order = sum(!is.na(birthorder_uterus_preg) & any_multiple_birth != 1,
                                           na.rm = T),
            Parental_Full_Sibling_Birth_Order = sum(!is.na(birthorder_genes) & any_multiple_birth != 1,
                                                    na.rm = T),
            Naive_Maternal_Birth_Order = sum(!is.na(birthorder_naive) & (is.na(any_multiple_birth) | any_multiple_birth != 1),
                                                    na.rm = T))
n_birthorder
```

### No single children
```{r}
## Count Birthorder - without only children
n_birthorder = alldata_birthorder %>%
  summarise(Maternal_Birth_Order = sum(!is.na(birthorder_uterus_alive) & any_multiple_birth != 1 &
                                         sibling_count_uterus_alive != 1,
                                       na.rm = T),
            Maternal_Pregnancy_Order = sum(!is.na(birthorder_uterus_preg) & any_multiple_birth != 1 &
                                             sibling_count_uterus_preg != 1,
                                           na.rm = T),
            Parental_Full_Sibling_Birth_Order = sum(!is.na(birthorder_genes) & any_multiple_birth != 1 &
                                                      sibling_count_genes != 1,
                                                    na.rm = T),
            Naive_Maternal_Birth_Order = sum(!is.na(birthorder_naive) & any_multiple_birth != 1 &
                                                      sibling_count_naive != 1,
                                                    na.rm = T))
n_birthorder
```


### Still alive
```{r}
## Count Birthorder - without only children
alldata_birthorder = alldata_birthorder %>%
  mutate(alive = ifelse(!is.na(g_factor_2015_old), 1, alive)) # for seven people its reported that they are not alive anymore but they participated in IFLS5 ?

n_birthorder = alldata_birthorder %>%
  summarise(Maternal_Birth_Order = sum(!is.na(birthorder_uterus_alive) & any_multiple_birth != 1 &
                                         sibling_count_uterus_alive != 1 & alive != 3,
                                       na.rm = T),
            Maternal_Pregnancy_Order = sum(!is.na(birthorder_uterus_preg) & any_multiple_birth != 1 &
                                             sibling_count_uterus_preg != 1 & alive != 3,
                                           na.rm = T),
            Parental_Full_Sibling_Birth_Order = sum(!is.na(birthorder_genes) & any_multiple_birth != 1 &
                                                      sibling_count_genes != 1 & alive != 3,
                                                    na.rm = T))
n_birthorder


```


### Interviewed for IFLS 5  {.tabset}

#### Total
```{r}
### All people in the birthorder data file that have ANY g-factor, educational attainment, personality or risk values have to be marked
alldata_birthorder = alldata_birthorder %>%
  mutate(used_any = ifelse(!is.na(g_factor_2015_old) | !is.na(g_factor_2015_young) | !is.na(g_factor_2007_old) |
                             !is.na(g_factor_2007_young) | !is.na(big5_ext) | !is.na(big5_open) | !is.na(big5_agree) |
                             !is.na(big5_neu) | !is.na(big5_con) | !is.na(riskA) | !is.na(riskB) |
                             is.na(years_of_education), 1, 0))

n_birthorder = alldata_birthorder %>%
  summarise(Maternal_Birth_Order = sum(!is.na(birthorder_uterus_alive) & any_multiple_birth != 1 &
                                         sibling_count_uterus_alive != 1 & alive != 3 &
                                         used_any == 1,
                                       na.rm = T),
            Maternal_Pregnancy_Order = sum(!is.na(birthorder_uterus_preg) & any_multiple_birth != 1 &
                                             sibling_count_uterus_preg != 1 & alive != 3 &
                                         used_any == 1,
                                           na.rm = T),
            Parental_Full_Sibling_Birth_Order = sum(!is.na(birthorder_genes) & any_multiple_birth != 1 &
                                                      sibling_count_genes != 1 & alive != 3 &
                                                      used_any == 1,
                                                    na.rm = T),
            Naive_Maternal_Birth_Order = sum(!is.na(birthorder_naive) & (is.na(any_multiple_birth) | any_multiple_birth != 1) &
                                                      sibling_count_naive != 1 & alive != 3 &
                                                      used_any == 1,
                                                    na.rm = T)

    )
n_birthorder
```

#### Intelligence
```{r}
### All people in the birthorder data file that have all Intelligence measurements
alldata_birthorder = alldata_birthorder %>%
  mutate(used_iq = ifelse(!is.na(g_factor_2015_old) | !is.na(g_factor_2015_young) | !is.na(g_factor_2007_old) |
                             !is.na(g_factor_2007_young), 1, 0)) # used_iq = 1:included, 0:not_included


n_birthorder = alldata_birthorder %>%
  summarise(Maternal_Birth_Order = sum(!is.na(birthorder_uterus_alive) & any_multiple_birth != 1 &
                                         sibling_count_uterus_alive != 1 & alive != 3 &
                                         used_iq == 1,
                                       na.rm = T),
            Maternal_Pregnancy_Order = sum(!is.na(birthorder_uterus_preg) & any_multiple_birth != 1 &
                                             sibling_count_uterus_preg != 1 & alive != 3 &
                                         used_iq == 1,
                                           na.rm = T),
            Parental_Full_Sibling_Birth_Order = sum(!is.na(birthorder_genes) & any_multiple_birth != 1 &
                                                      sibling_count_genes != 1 & alive != 3 &
                                                      used_iq == 1,
                                                    na.rm = T))
n_birthorder

```


#### Personality
```{r}
### All people in the birthorder data file that have Personality Measurements
missingness_patterns(alldata_birthorder %>% select(big5_ext, big5_con, big5_open, big5_neu, big5_agree, big5_open))
# Missingness Pattern shows that if one of the personality measurements is missing all Measurements are missing
alldata_birthorder = alldata_birthorder %>%
  mutate(used_pers = ifelse(!is.na(big5_ext), 1, 0))


n_birthorder = alldata_birthorder %>%
  summarise(Maternal_Birth_Order = sum(!is.na(birthorder_uterus_alive) & any_multiple_birth != 1 &
                                         sibling_count_uterus_alive != 1 & alive != 3 &
                                         used_pers == 1,
                                       na.rm = T),
            Maternal_Pregnancy_Order = sum(!is.na(birthorder_uterus_preg) & any_multiple_birth != 1 &
                                             sibling_count_uterus_preg != 1 & alive != 3 &
                                         used_pers == 1,
                                           na.rm = T),
            Parental_Full_Sibling_Birth_Order = sum(!is.na(birthorder_genes) & any_multiple_birth != 1 &
                                                      sibling_count_genes != 1 & alive != 3 &
                                                      used_pers == 1,
                                                    na.rm = T))
n_birthorder

```

#### Risk preference
```{r}
### All people in the birthorder data file that have Risk Measurements
missingness_patterns(alldata_birthorder %>% select(riskA, riskB))
## Missingness Pattern shows that sometimes A and sometimes B is missing
alldata_birthorder = alldata_birthorder %>%
  mutate(used_riskA = ifelse(!is.na(riskA), 1, 0),
         used_riskB = ifelse(!is.na(riskB), 1, 0))

n_birthorder = alldata_birthorder %>%
  summarise(Maternal_Birth_Order = sum(!is.na(birthorder_uterus_alive) & any_multiple_birth != 1 &
                                         sibling_count_uterus_alive != 1 & alive != 3 &
                                         used_riskA == 1,
                                       na.rm = T),
            Maternal_Pregnancy_Order = sum(!is.na(birthorder_uterus_preg) & any_multiple_birth != 1 &
                                             sibling_count_uterus_preg != 1 & alive != 3 &
                                         used_riskA == 1,
                                           na.rm = T),
            Parental_Full_Sibling_Birth_Order = sum(!is.na(birthorder_genes) & any_multiple_birth != 1 &
                                                      sibling_count_genes != 1 & alive != 3 &
                                                      used_riskA == 1,
                                                    na.rm = T))
n_birthorder



n_birthorder = alldata_birthorder %>%
  summarise(Maternal_Birth_Order = sum(!is.na(birthorder_uterus_alive) & any_multiple_birth != 1 &
                                         sibling_count_uterus_alive != 1 & alive != 3 &
                                         used_riskB == 1,
                                       na.rm = T),
            Maternal_Pregnancy_Order = sum(!is.na(birthorder_uterus_preg) & any_multiple_birth != 1 &
                                             sibling_count_uterus_preg != 1 & alive != 3 &
                                         used_riskB == 1,
                                           na.rm = T),
            Parental_Full_Sibling_Birth_Order = sum(!is.na(birthorder_genes) & any_multiple_birth != 1 &
                                                      sibling_count_genes != 1 & alive != 3 &
                                                      used_riskB == 1,
                                                    na.rm = T))
n_birthorder


```

#### Educational Attainment
```{r}
### All people in the birthorder data file that have educational attainment measurements
missingness_patterns(alldata_birthorder %>% select(years_of_education))

alldata_birthorder = alldata_birthorder %>%
  mutate(used_ea = ifelse(!is.na(years_of_education), 1, 0))

n_birthorder = alldata_birthorder %>%
  summarise(Maternal_Birth_Order = sum(!is.na(birthorder_uterus_alive) & any_multiple_birth != 1 &
                                         sibling_count_uterus_alive != 1 & alive != 3 &
                                         used_ea == 1,
                                       na.rm = T),
            Maternal_Pregnancy_Order = sum(!is.na(birthorder_uterus_preg) & any_multiple_birth != 1 &
                                             sibling_count_uterus_preg != 1 & alive != 3 &
                                         used_ea == 1,
                                           na.rm = T),
            Parental_Full_Sibling_Birth_Order = sum(!is.na(birthorder_genes) & any_multiple_birth != 1 &
                                                      sibling_count_genes != 1 & alive != 3 &
                                                      used_ea == 1,
                                                    na.rm = T))
n_birthorder

```
