==> 3_imputation.Rmd <==
---
title: "3_imputation"
author: "Laura Botzet & Ruben Arslan"
date: "28 September 2017"
output: 
  html_document:
    code_folding: "show"
editor_options: 
  chunk_output_type: console
---

# <span style="color:#FFD92F">Imputation</span> {.tabset}

## Helper
```{r helper}
knitr::opts_chunk$set(eval = T)
source("0_helpers.R")
library(tidylog)
```


## Data
```{r}
### Import alldata used for birthorder and g-factors computed
alldata_birthorder = readRDS("data/alldata_birthorder.rds")

## For the imputation we want to clean the dataset and get rid of all uninteresting data
alldata_birthorder = alldata_birthorder %>%
  filter(!is.na(pidlink)) %>% # no individuals who are only known from the pregnancy file
  filter(is.na(lifebirths) | lifebirths == 2) %>% # remove 7 and 2 individuals who are known as stillbirth or miscarriage but still have PID
  select(-lifebirths) %>% 
  filter(!is.na(mother_pidlink)) %>% 
  select(-father_pidlink) %>% 
  filter(is.na(any_multiple_birth) | any_multiple_birth != 1) %>% # remove 7 and 2 individuals who are known as stillbirth or miscarriage but still have PID %>% 
  select(-starts_with("age_"), -wave, -any_multiple_birth, -multiple_birth,
         -starts_with("g_factor")) %>% 
  mutate(money_spent_smoking_log = if_else(is.na(money_spent_smoking_log) & ever_smoked == 0, 0, money_spent_smoking_log),
         amount = if_else(is.na(amount) & ever_smoked == 0, 0, amount),
         amount_still_smokers = if_else(is.na(amount_still_smokers) &  still_smoking == 0, 0, amount_still_smokers),
         birthyear = lubridate::year(birthdate))

## For the second imputation we only want to keep nice, non-problematic variables
alldata_birthorder = alldata_birthorder %>%
  select(-birthorder_uterus_alive, -birthorder_uterus_preg,
         -sibling_count_uterus_preg, -sibling_count_naive_ind, -sibling_count_uterus_alive,
         -words_immediate, -words_remembered_avg,
         -Total_score_highest,-Total_score_highest_type, -Math_score_highest,
         -Math_score_highest_type, -birthdate, -birthyear, -marriage_id,
         -money_spent_smoking_log,-amount_still_smokers, -total_worked, -total_missed, 
         -highest_education, -chron_order_birth, -alive, -death_yr, -death_month,
         -University_worked, -University_missed, - English_score_elementary,
         )

alldata_birthorder = alldata_birthorder %>%
  mutate(currently_attending_school = as.integer(currently_attending_school),
         attended_school = as.integer(attended_school),
         Type_of_test_elementary = as.integer(Type_of_test_elementary),
         Type_of_test_Junior_High = as.integer(Type_of_test_Junior_High),
         Type_of_test_Senior_High = as.integer(Type_of_test_Senior_High),
         random_si = as.integer(random_si)
         )

# recode Factor Variable as Dummy Variable
alldata_birthorder = left_join(alldata_birthorder,
                                alldata_birthorder %>%
                                  filter(!is.na(Category)) %>%
                                  mutate(var = 1) %>%
                                  select(pidlink, Category, var) %>%
                                  spread(Category, var, fill = 0, sep = "_"), by = "pidlink") %>%
  select(-Category)

# recode Factor Variable as Dummy Variable
alldata_birthorder = left_join(alldata_birthorder,
                                alldata_birthorder %>%
                                  filter(!is.na(Sector)) %>%
                                  mutate(var = 1) %>%
                                  select(pidlink, Sector, var) %>%
                                  spread(Sector, var, fill = 0, sep = "_"), by = "pidlink") %>%
  select(-Sector)



alldata_birthorder = alldata_birthorder %>% 
  mutate(
# birthorder as factors with levels of 1, 2, 3, 4, 5, 5+
    birthorder_naive_factor = as.character(birthorder_naive),
    birthorder_naive_factor = ifelse(birthorder_naive > 5, "5+",
                                            birthorder_naive_factor),
    birthorder_naive_factor = factor(birthorder_naive_factor, 
                                            levels = c("1","2","3","4","5","5+")),
    sibling_count_naive_factor = as.character(sibling_count_naive),
    sibling_count_naive_factor = ifelse(sibling_count_naive > 5, "5+",
                                               sibling_count_naive_factor),
    sibling_count_naive_factor = factor(sibling_count_naive_factor, 
                                               levels = c("2","3","4","5","5+")),
    birthorder_genes_factor = as.character(birthorder_genes),
    birthorder_genes_factor = ifelse(birthorder_genes >5 , "5+", birthorder_genes_factor),
    birthorder_genes_factor = factor(birthorder_genes_factor, 
                                     levels = c("1","2","3","4","5","5+")),
    sibling_count_genes_factor = as.character(sibling_count_genes),
    sibling_count_genes_factor = ifelse(sibling_count_genes >5 , "5+",
                                        sibling_count_genes_factor),
    sibling_count_genes_factor = factor(sibling_count_genes_factor, 
                                        levels = c("2","3","4","5","5+")),
    # interaction birthorder * siblingcout for each birthorder
    count_birthorder_naive =
      factor(str_replace(as.character(interaction(birthorder_naive_factor,                                                              sibling_count_naive_factor)),
                        "\\.", "/"),
                                           levels =   c("1/2","2/2", "1/3",  "2/3",
                                                        "3/3", "1/4", "2/4", "3/4", "4/4",
                                                        "1/5", "2/5", "3/5", "4/5", "5/5",
                                                        "1/5+", "2/5+", "3/5+", "4/5+",
                                                        "5/5+", "5+/5+")),
    count_birthorder_genes =
      factor(str_replace(as.character(interaction(birthorder_genes_factor,                                                              sibling_count_genes_factor)), "\\.", "/"),
                                           levels =   c("1/2","2/2", "1/3",  "2/3",
                                                        "3/3", "1/4", "2/4", "3/4", "4/4",
                                                        "1/5", "2/5", "3/5", "4/5", "5/5",
                                                        "1/5+", "2/5+", "3/5+", "4/5+",
                                                        "5/5+", "5+/5+"))) %>% 
  select(-sibling_count_genes_factor, -sibling_count_naive_factor)


corr_pipe = . %>%  as.matrix() %>% 
Hmisc::rcorr(.) %>% {
  left_join(left_join(.$r %>% reshape2::melt() %>% rename(r = value), 
                      .$n %>% reshape2::melt() %>% rename(n = value), by = c("Var1", "Var2")),
            .$P %>% reshape2::melt() %>% rename(p = value), 
            by = c("Var1", "Var2"))
  } %>% 
  arrange(desc(r)) %>% 
  filter(Var1 != Var2)

attributes(alldata_birthorder$mother_pidlink)
zap_label <- function(x) {
  UseMethod("zap_label")
}
zap_label.data.frame <- function(x) {
  x[] <- lapply(x, zap_label)
  x
}
zap_label.default <- function(x) {
  attributes(x)$label <- NULL
  x
}
  
alldata_birthorder <- alldata_birthorder %>% haven::zap_labels() %>% haven::zap_formats() %>% zap_label()
str(alldata_birthorder)
alldata_birthorder %>% select_if(function(.)!is.numeric(.)) %>% names()

correlations = corr_pipe(alldata_birthorder %>% select_if(is.numeric))
```

## Defining the sample for imputation
```{r}
alldata_birthorder$main_outcomes <- alldata_birthorder %>% select(`raven_2015_young`, `raven_2015_old`, `words_delayed`, `raven_2007_old`, `raven_2007_young`, `big5_ext`, `riskA`, `years_of_education`) %>% mutate_all(~ !is.na(.)) %>% rowSums()
xtabs(~ main_outcomes + age, alldata_birthorder)
# xtabs(~ is.na(raven_2015_young) + age, alldata_birthorder)
alldata_birthorder <-  alldata_birthorder %>% 
  # filter(!is.na(birthorder_genes)) %>% 
  group_by(mother_pidlink) %>% 
  mutate(var_scg = var(sibling_count_genes, na.rm = T)) %>% 
  filter(is.na(var_scg) | var_scg == 0) %>% 
  ungroup() %>% 
  filter(coalesce(sibling_count_genes, sibling_count_naive) > 1) %>% 
  filter(!is.na(birthorder_genes) | ( !is.na(birthorder_naive) & main_outcomes > 1)) %>% 
  filter(age > 14 | (age > 13) & (!is.na(raven_2015_old) | !is.na(years_of_education) | !is.na(big5_ext) | !is.na(riskA))) %>% 
  select(`mother_pidlink`,  `age`, `male`, 
         `birthorder_naive`, `sibling_count_naive`,  count_birthorder_naive,
         `birthorder_genes`, `sibling_count_genes`,  count_birthorder_genes,
`raven_2015_young`, `math_2015_young`, `raven_2015_old`, `math_2015_old`, `words_delayed`, `adaptive_numbering`, `raven_2007_old`, `raven_2007_young`, `math_2007_young`, `math_2007_old`, `count_backwards`, 

`big5_ext`, `big5_con`, `big5_open`, `big5_neu`, `big5_agree`, `riskA`, `riskB`,

`years_of_education`)
  
alldata_birthorder <-  alldata_birthorder %>% 
  group_by(mother_pidlink) %>% 
  mutate(sibling_count_genes = if_else(all(is.na(sibling_count_genes)), NA_integer_, as.integer(mean(sibling_count_genes, na.rm = T)))) %>% 
  ungroup()

alldata_birthorder <-  alldata_birthorder %>% 
  rename_all(funs(str_replace_all(., "\\s", "_"))) %>% 
  rename_all(funs(str_replace_all(., ",", "_")))

age_poly <- poly(scale(alldata_birthorder$age)[,1], degree = 3, raw = T)
qplot(age_poly[,1], alldata_birthorder$age)

age_groups <- Hmisc::cut2(alldata_birthorder$age, cuts = c(14, 18, 25, 30, 40, 53))
table(as.character(age_groups), exclude = NULL)
# alldata_birthorder$age <- age_groups
alldata_birthorder <- alldata_birthorder %>%
  mutate(
    age = age_poly[,1],
    age2 = age_poly[,2],
    age3 = age_poly[,3])

saveRDS(alldata_birthorder, "data/alldata_birthorder_missing.rds")
```

## Functions

```{r}
library(mice)
library(miceadds)
library(mitml)
library(lattice)
library(pan)
# skimr::skim(alldata_birthorder)
```


## Imputation {.active .tabset}
```{r}
set.seed(123) # random seed so that other can replicate our imputation

# look at the dataset
colnames(alldata_birthorder)
head(alldata_birthorder)
dim(alldata_birthorder)
nrow(alldata_birthorder)
ncol(alldata_birthorder)
summary(alldata_birthorder)
skimr::skim(alldata_birthorder)
```


### Missingness pattern

We can see that 

1. there are many for whom we only have birth order information, no outcomes
2. there are many for whom we have outcome info, but only the "naive" birth order, not based
on pregnancy data

#### Reduced set of variables to highlight patterns

```{r}
sort(round(miss_frac(alldata_birthorder),2))
missingness_pattern = md.pattern(alldata_birthorder %>% 
                                   select(birthorder_genes,
                                          math_2015_young,
                                          math_2015_old, words_delayed,
                                          raven_2007_young, raven_2007_old, big5_ext))
DT::datatable(md_pattern(alldata_birthorder %>% 
                                   select(birthorder_genes,
                                          math_2015_young,
                                          math_2015_old, words_delayed,
                                          raven_2007_young, raven_2007_old, big5_ext), min_freq = 0.005))
```

#### All variables
```{r}
#look at the missing data pattern
missingness_pattern = md.pattern(alldata_birthorder)
DT::datatable(md_pattern(alldata_birthorder, min_freq = 0.005))
```

### prepare Imputation
```{r prepimp}
alldata_birthorder$mother_pidlink = as.integer(alldata_birthorder$mother_pidlink)

imp0 <- mice(alldata_birthorder, maxit=0)

pred1 <- imp0$predictorMatrix
pred1
pred1 = quickpred(alldata_birthorder, mincor = 0.05) #all variables that correlate > .05
pred1
# pred1[, 1:27] <- 1 # let focal variables predict each other in any case

# naive BO/SC never imputed (almost never missing, only used to impute BO genes)
pred1[, c("birthorder_naive", "sibling_count_naive", "count_birthorder_naive")] <- 0
# naive BO/SC never used to impute traits
pred1[c("birthorder_naive", "sibling_count_naive", "count_birthorder_naive"), ] <- 0

# linear BO/SC genes not used to impute traits, only their interaction count_birthorder_genes
pred1[c("birthorder_genes", "sibling_count_genes"), ] <- 0

# naive BO only used to impute BO genes
pred1[, c("birthorder_genes", "sibling_count_genes", "count_birthorder_genes")] <- 0
pred1["birthorder_naive", c("birthorder_genes")] <- 1
pred1["sibling_count_naive", c("sibling_count_genes")] <- 1
pred1["count_birthorder_naive", c("count_birthorder_genes")] <- 1
pred1["count_birthorder_genes", c("raven_2015_young", "math_2015_young", 
"raven_2015_old", "math_2015_old", "words_delayed", "adaptive_numbering", 
"raven_2007_old", "raven_2007_young", "math_2007_young", "math_2007_old", 
"count_backwards", "big5_ext", "big5_con", "big5_open", "big5_neu", 
"big5_agree", "riskA", "riskB", "years_of_education")] <- 1

meth1 <- imp0$method

# linear multilevel imputation using the function mice.impute.2l.norm()
# specify predictor_matrix using 1 = fixed effects, 2 = random effects, -2 = class variable, 0 = no effect
# we can set the mother_pidlink as class variable (= -2)
pred1[, "mother_pidlink"] = -2
# set imputation procedures
meth1[-1] <- "2l.pan"
meth1[c("birthorder_naive")] <- ""
meth1[c("sibling_count_naive")] <- ""
meth1[c("count_birthorder_naive")] <- ""
meth1[c("birthorder_genes")] <- "cart"
meth1[c("sibling_count_genes")] <- "2lonly.norm"
meth1[c("count_birthorder_genes")] <- "polyreg"
pred1
meth1
```


### m=1
```{r}
# imp <- mice(alldata_birthorder, m = 1, maxit = 1, predictorMatrix = pred1, method = meth1)
# mids2mitml.list(imp)[[1]]$sibling_count_genes %>% table(exclude = NULL)
# mids2mitml.list(imp)[[1]] %>% group_by(mother_pidlink) %>% summarise(n_distinct(birthorder_genes),
#                                                     n_distinct(sibling_count_genes),
#                                                     n(), 
#                                                     mean(sibling_count_genes, na.rm =T)) %>% 
#   View()

# imp$loggedEvents
# densityplot(imp)
```


### Run more imputations on cluster
```{r clusterimp}
if (file.exists("data/birthorder_imps4.rds")) {
  all_imps <- readRDS("data/birthorder_imps4.rds")
} else {
  # imp1 <- mice(alldata_birthorder, maxit = 1, m=1, predictorMatrix = pred1, method = meth1)
  # imp2 <- mice(alldata_birthorder, maxit = 60, m=1, predictorMatrix = pred1, method = meth1)
  
  library(future)
library(future.batchtools)

#' # Send job to cluster {.tabset .tabset-sticky}
#' This function is used to send jobs to a computing cluster.
#' It accomplishes the following needs:
#' - wait for errors during building the model
#' - close the SSH connections to the cluster if building the model worked
#'    - this is useful so I don't open hundreds of SSH tunnels to send over
#'      models
#' - if the model has been built already, check if LOO has been computed, if not, rerun
#' - if we're not in an interactive session, load the model (this allows me to use
#' the same code to retrieve the model that I use to send it off)
#'
send_job_to_cluster <- function(expr,
                                job_name = 'boimp', queue = "default",
                                walltime = "10:0:0", memory = 10, ncpus = 6,
                                interactive_call = interactive(),
                                wait_till_error = 3,
                                template = 'torque-lido.tmpl') {
  file <- paste0(job_name, ".rds")
  file_there <- file.exists(file)
  if (file_there) {
    mod <- readRDS(file)
    if (interactive_call) {
      "done"
    } else {
      mod
    }
  } else {
    if (file.exists(template)) { 
      qsub <- tweak(batchtools_torque, template = template,
                    # workers = "export LSF_ENVDIR=/opt/lsf/conf",
                    resources = list(job.name = job_name,
                                     queue = queue,
                                     walltime = walltime,
                                     memory = memory,
                                     ncpus = ncpus))
      plan(qsub)
    }
    expr <- substitute(expr)
    job <- future(expr, substitute = FALSE,
                  conditions = character(0),
                  stdout = NA)
    i <- 0
    while (i < wait_till_error) {
      if (resolved(job)) {
        break
      } else {
        Sys.sleep(3)
        i <- i + 3
      }
    }
    if (resolved(job)) {
      value(job)
    } else {
      "sent"
    }
  }
}
model_time <- function(model) {
  rstan::get_elapsed_time(model$fit) %>% rowSums() %>% max() %>% {./60/60}
}

options(future.globals.maxSize = 2000*1024^2)
# login just once, deliver multiple jobs
login <- tweak(remote, workers = rep("arslan@tardis.mpib-berlin.mpg.de", 1),
               persistent = FALSE)
plan(login)

  library(future)
  m <- 50
  all_imps %<-% {
    imps <- listenv::listenv()
    for (i in seq_along(1:m)) {
      job_name <- paste0("bo_imp/i6", i)
      imps[[i]] <- send_job_to_cluster({
        saveRDS(1, file = paste0(job_name, ".rds"))
        imp <- mice(alldata_birthorder, m = 1, maxit = 30, predictorMatrix = pred1, method = meth1)
        saveRDS(imp, file = paste0(job_name, ".rds"))
        imp
      }, job_name = job_name, queue = "default", ncpus = 1, walltime = "5:0:0",
memory = 2)
    }
    all_imps <- as.list(imps)
    # saveRDS(all_imps, "birthorder_imps4.rds")
    all_imps
  }
  saveRDS(all_imps, "data/birthorder_imps4.rds")
}

alldata_birthorder_i = all_imps[[1]]
for (i in 2:length(all_imps)) {
  alldata_birthorder_i = ibind(alldata_birthorder_i, all_imps[[i]])
}
```



```{r}
saveRDS(alldata_birthorder_i, file = "data/alldata_birthorder_i_ml5.rds")
```
