==> 3_sample_analyses.Rmd <==
---
title: "3_sample_analyses"
author: "Laura Botzet & Ruben Arslan"
output: html_document
editor_options: 
  chunk_output_type: console
---
#  <span style="color:#E78AC3">Sample Analyses</span> {.tabset}

## Helper
```{r helper}
source("0_helpers.R")
```

## Load data
```{r Load data}
### Import all data with known birthorder
birthorder = readRDS("data/alldata_birthorder.rds")
```

## Data wrangling
```{r data wrangling}
## we have to exclude people in the control group who are part of the birthorder group
birthorder = birthorder %>%
  # mark people who have missing birthorder data
  mutate(check_birthorder = ifelse(!is.na(birthorder_genes), 1, 0),
  # mark people who have missing outcomes
         check_outcome = ifelse(!is.na(raven_2015_old), 1,
                         ifelse(!is.na(math_2015_old), 1,
                         ifelse(!is.na(raven_2015_young), 1,
                         ifelse(!is.na(math_2015_young), 1,
                         ifelse(!is.na(raven_2007_old), 1,
                         ifelse(!is.na(math_2007_old), 1,
                         ifelse(!is.na(raven_2007_young), 1,
                         ifelse(!is.na(math_2007_young), 1,
                         ifelse(!is.na(adaptive_numbering), 1,
                         ifelse(!is.na(words_remembered_avg), 1,
                         ifelse(!is.na(count_backwards), 1,
                         ifelse(!is.na(big5_ext), 1, 
                         ifelse(!is.na(riskA), 1,
                         ifelse(!is.na(riskB), 1,
                         ifelse(!is.na(years_of_education), 1,
                         ifelse(!is.na(Elementary_missed), 1,
                         ifelse(!is.na(Elementary_worked), 1,
                         ifelse(!is.na(attended_school), 1,
                         ifelse(!is.na(wage_last_month_log), 1,
                         ifelse(!is.na(wage_last_year_log), 1,
                         ifelse(!is.na(Self_employed), 1,
                         ifelse(!is.na(Category), 1,
                         ifelse(!is.na(Sector), 1,
                         ifelse(!is.na(ever_smoked), 1,
                         ifelse(!is.na(still_smoking), 1,
                         0))))))))))))))))))))))))),
  group_birthorder = ifelse(check_birthorder == 0, "control", "test"),
  group_outcome = ifelse(check_outcome == 0, "control", "test"))

```

## Effects of Educational Attainment and Intelligence on Selection
```{r}
library(sjstats)
library(arm)

birthorder = birthorder %>%
  mutate(main_sample = ifelse(group_birthorder == "test" & group_outcome == "test", 1, 0))

glm(main_sample ~ g_factor_2015_old + years_of_education,
    data = birthorder,
    family = binomial(link='logit'))
std_beta(glm(main_sample ~ g_factor_2015_old + years_of_education, data = birthorder),
    family = binomial(link='logit'), ci.lvl = 0.995)

glm(main_sample ~ g_factor_2015_old + years_of_education +
      poly(age, 3, raw = TRUE) + sibling_count_naive,
    data = birthorder,
    family = binomial(link='logit'),
    control=glm.control(maxit=50))
std_beta(glm(main_sample ~ g_factor_2015_old + years_of_education +
               poly(age, 3, raw = TRUE) + sibling_count_naive,
             data = birthorder),
         family = binomial(link='logit'), ci.lvl = 0.995)


glm(main_sample ~ g_factor_2015_old + years_of_education +
      poly(age, 3, raw = TRUE) + sibling_count_naive + male,
    data = birthorder,
    family = binomial(link='logit'),
    control=glm.control(maxit=50))
std_beta(glm(main_sample ~ g_factor_2015_old + years_of_education +
               poly(age, 3, raw = TRUE) + sibling_count_naive + male,
             data = birthorder),
         family = binomial(link='logit'), ci.lvl = 0.995)

glm(main_sample ~ g_factor_2015_old,
    data = birthorder,
    family = binomial(link='logit'))
std_beta(glm(main_sample ~ g_factor_2015_old, data = birthorder),
    family = binomial(link='logit'), ci.lvl = 0.995)

glm(main_sample ~ g_factor_2015_old +
      poly(age, 3, raw = TRUE) + sibling_count_naive,
    data = birthorder,
    family = binomial(link='logit'),
    control=glm.control(maxit=50))
std_beta(glm(main_sample ~ g_factor_2015_old +
               poly(age, 3, raw = TRUE) + sibling_count_naive,
             data = birthorder),
         family = binomial(link='logit'), ci.lvl = 0.995)

glm(main_sample ~ years_of_education,
    data = birthorder,
    family = binomial(link='logit'))
std_beta(glm(main_sample ~ years_of_education, data = birthorder),
    family = binomial(link='logit'), ci.lvl = 0.995)

glm(main_sample ~ years_of_education +
      poly(age, 3, raw = TRUE) + sibling_count_naive,
    data = birthorder,
    family = binomial(link='logit'),
    control=glm.control(maxit=50))
std_beta(glm(main_sample ~ years_of_education +
               poly(age, 3, raw = TRUE) + sibling_count_naive,
             data = birthorder),
         family = binomial(link='logit'), ci.lvl = 0.995)

```

## Birth order and sibship size {.active .tabset}

### Genes birthorder
```{r maternal birthorder}
descriptives = birthorder %>%
  group_by(group_outcome) %>%
  summarise(individuals = n(),
            mothers = length(unique(mother_pidlink)),
            sibship_size_mean = mean(sibling_count_genes, na.rm = T),
            sibship_size_confidence_low = mean(sibling_count_genes, na.rm = T) - 
              (qt(.975, n()-1)*sd(sibling_count_genes, na.rm = T)/sqrt(n())),
            sibship_size_confidence_high = mean(sibling_count_genes, na.rm = T) + 
              (qt(.975, n()-1)*sd(sibling_count_genes, na.rm = T)/sqrt(n())),
            sibship_size_min = min(sibling_count_genes, na.rm = TRUE),
            sibship_size_max = max(sibling_count_genes, na.rm = TRUE),
            birthorder_mean = mean(birthorder_genes, na.rm = T),
            birthorder_size_confidence_low = mean(birthorder_genes, na.rm = T) - 
              (qt(.975, n()-1)*sd(birthorder_genes, na.rm = T)/sqrt(n())),
            birthorder_size_confidence_high = mean(birthorder_genes, na.rm = T) + 
              (qt(.975, n()-1)*sd(birthorder_genes, na.rm = T)/sqrt(n())),
            birthorder_size_min = min(birthorder_genes, na.rm = TRUE),
            birthorder_size_max = max(birthorder_genes, na.rm = TRUE),
            number_siblings_mean = mean(sibling_count_genes, na.rm =T),
            number_siblings_confidence_low = mean(sibling_count_genes, na.rm = T) -
              (qt(.975, n()-1)*sd(sibling_count_genes, na.rm = T)/sqrt(n())),
            number_siblings_confidence_high = mean(sibling_count_genes, na.rm = T) +
              (qt(.975, n()-1)*sd(sibling_count_genes, na.rm = T)/sqrt(n())))

descriptives

birthorder %>% 
  t.test(sibling_count_genes ~ group_outcome, data = ., var.equal = T)

birthorder %>% 
  t.test(birthorder_genes ~ group_outcome, data = ., var.equal = T)

```


#### How many siblings with data do we retain in each family?
```{r}
birthorder %>% filter(!is.na(birthorder_genes), group_outcome == "test") %>% group_by(mother_pidlink) %>% summarise(with_data = n(), all = mean(sibling_count_genes)) -> counts
birthorder %>% filter(!is.na(birthorder_genes), group_outcome == "control") %>% group_by(mother_pidlink) %>% summarise(with_data = n(), all = mean(sibling_count_genes)) -> counts1
```

In our test sample families with an average size of `r mean(counts$all,na.rm= T)` siblings, we retain `r mean(counts$with_data)`.

In the original sample families with an average size of `r mean(counts1$all,na.rm= T)` siblings, we retain `r mean(counts1$with_data)`.

```{r}
ggplot(counts, aes(all, with_data)) + geom_jitter(alpha = 0.1) + geom_smooth() + scale_x_continuous(breaks=1:15) + scale_y_continuous(breaks=1:15)
```

### Maternal birthorder
```{r maternal birthorder}
descriptives = birthorder %>%
  group_by(group_outcome) %>%
  summarise(individuals = n(),
            mothers = length(unique(mother_pidlink)),
            sibship_size_mean = mean(sibling_count_uterus_alive, na.rm = T),
            sibship_size_confidence_low = mean(sibling_count_uterus_alive, na.rm = T) - 
              (qt(.975, n()-1)*sd(sibling_count_uterus_alive, na.rm = T)/sqrt(n())),
            sibship_size_confidence_high = mean(sibling_count_uterus_alive, na.rm = T) + 
              (qt(.975, n()-1)*sd(sibling_count_uterus_alive, na.rm = T)/sqrt(n())),
            sibship_size_min = min(sibling_count_uterus_alive, na.rm = TRUE),
            sibship_size_max = max(sibling_count_uterus_alive, na.rm = TRUE),
            birthorder_mean = mean(birthorder_uterus_alive, na.rm = T),
            birthorder_size_confidence_low = mean(birthorder_uterus_alive, na.rm = T) - 
              (qt(.975, n()-1)*sd(birthorder_uterus_alive, na.rm = T)/sqrt(n())),
            birthorder_size_confidence_high = mean(birthorder_uterus_alive, na.rm = T) + 
              (qt(.975, n()-1)*sd(birthorder_uterus_alive, na.rm = T)/sqrt(n())),
            birthorder_size_min = min(birthorder_uterus_alive, na.rm = TRUE),
            birthorder_size_max = max(birthorder_uterus_alive, na.rm = TRUE),
            number_siblings_mean = mean(sibling_count_uterus_alive, na.rm =T),
            number_siblings_confidence_low = mean(sibling_count_uterus_alive, na.rm = T) -
              (qt(.975, n()-1)*sd(sibling_count_uterus_alive, na.rm = T)/sqrt(n())),
            number_siblings_confidence_high = mean(sibling_count_uterus_alive, na.rm = T) +
              (qt(.975, n()-1)*sd(sibling_count_uterus_alive, na.rm = T)/sqrt(n())))

descriptives

birthorder %>% 
  t.test(sibling_count_uterus_alive ~ group_outcome, data = ., var.equal = T)

birthorder %>% 
  t.test(birthorder_uterus_alive ~ group_outcome, data = ., var.equal = T)

```


#### How many siblings with data do we retain in each family?
```{r}

birthorder %>% filter(!is.na(birthorder_uterus_alive), group_outcome == 1) %>% group_by(mother_pidlink) %>% summarise(with_data = n(), all = mean(sibling_count_uterus_alive)) -> counts
birthorder %>% filter(!is.na(birthorder_uterus_alive), group_outcome == 0) %>% group_by(mother_pidlink) %>% summarise(with_data = n(), all = mean(sibling_count_uterus_alive)) -> counts1
```

In our test sample families with an average size of `r mean(counts$all,na.rm= T)` siblings, we retain `r mean(counts$with_data)`.

In the original sample families with an average size of `r mean(counts1$all,na.rm= T)` siblings, we retain `r mean(counts1$with_data)`.

```{r}
ggplot(counts, aes(all, with_data)) + geom_jitter(alpha = 0.1) + geom_smooth() + scale_x_continuous(breaks=1:15) + scale_y_continuous(breaks=1:15)
```
### Maternal pregnancy birthorder
```{r maternal birthorder}
descriptives = birthorder %>%
  group_by(group_outcome) %>%
  summarise(individuals = n(),
            mothers = length(unique(mother_pidlink)),
            sibship_size_mean = mean(sibling_count_uterus_preg, na.rm = T),
            sibship_size_confidence_low = mean(sibling_count_uterus_preg, na.rm = T) - 
              (qt(.975, n()-1)*sd(sibling_count_uterus_preg, na.rm = T)/sqrt(n())),
            sibship_size_confidence_high = mean(sibling_count_uterus_preg, na.rm = T) + 
              (qt(.975, n()-1)*sd(sibling_count_uterus_preg, na.rm = T)/sqrt(n())),
            sibship_size_min = min(sibling_count_uterus_preg, na.rm = TRUE),
            sibship_size_max = max(sibling_count_uterus_preg, na.rm = TRUE),
            birthorder_mean = mean(birthorder_uterus_preg, na.rm = T),
            birthorder_size_confidence_low = mean(birthorder_uterus_preg, na.rm = T) - 
              (qt(.975, n()-1)*sd(birthorder_uterus_preg, na.rm = T)/sqrt(n())),
            birthorder_size_confidence_high = mean(birthorder_uterus_preg, na.rm = T) + 
              (qt(.975, n()-1)*sd(birthorder_uterus_preg, na.rm = T)/sqrt(n())),
            birthorder_size_min = min(birthorder_uterus_preg, na.rm = TRUE),
            birthorder_size_max = max(birthorder_uterus_preg, na.rm = TRUE),
            number_siblings_mean = mean(sibling_count_uterus_preg, na.rm =T),
            number_siblings_confidence_low = mean(sibling_count_uterus_preg, na.rm = T) -
              (qt(.975, n()-1)*sd(sibling_count_uterus_preg, na.rm = T)/sqrt(n())),
            number_siblings_confidence_high = mean(sibling_count_uterus_preg, na.rm = T) +
              (qt(.975, n()-1)*sd(sibling_count_uterus_preg, na.rm = T)/sqrt(n())))

descriptives

birthorder %>% 
  t.test(sibling_count_uterus_preg ~ group_outcome, data = ., var.equal = T)

birthorder %>% 
  t.test(birthorder_uterus_preg ~ group_outcome, data = ., var.equal = T)

```


#### How many siblings with data do we retain in each family?
```{r}
birthorder %>% filter(!is.na(birthorder_uterus_preg), group_outcome == 1) %>% group_by(mother_pidlink) %>% summarise(with_data = n(), all = mean(sibling_count_uterus_preg)) -> counts
birthorder %>% filter(!is.na(birthorder_uterus_preg), group_outcome == 0) %>% group_by(mother_pidlink) %>% summarise(with_data = n(), all = mean(sibling_count_uterus_preg)) -> counts1
```

In our test sample families with an average size of `r mean(counts$all,na.rm= T)` siblings, we retain `r mean(counts$with_data)`.

In the original sample families with an average size of `r mean(counts1$all,na.rm= T)` siblings, we retain `r mean(counts1$with_data)`.

```{r}
ggplot(counts, aes(all, with_data)) + geom_jitter(alpha = 0.1) + geom_smooth() + scale_x_continuous(breaks=1:15) + scale_y_continuous(breaks=1:15)
```


## Outcome measurements and covariates {.tabset}
### Control group
```{r data comparison}
## Descriptives
descriptives = birthorder %>%
  group_by(group_birthorder) %>%
    summarise(n = n(),
              age_mean = mean(age, na.rm=TRUE),
              age_confidence_low = mean(age, na.rm = T) - 
                (qt(.975, n()-1)*sd(age, na.rm = T)/sqrt(n())),
              age_confidence_high = mean(age, na.rm = T) + 
                (qt(.975, n()-1)*sd(age, na.rm = T)/sqrt(n())),
              age_min = min(age, na.rm = TRUE),
              age_max = max(age, na.rm = TRUE),
              gender = mean(male, na.rm=TRUE))
descriptives



## Ttest
tidy(t.test(birthorder$age ~ birthorder$group_birthorder, var.equal = T))

cohen.d(birthorder$age, as.factor(birthorder$group_birthorder), na.rm = T)

gender = birthorder %>%
  group_by(group_birthorder) %>%
    summarise(gender = sum(male == 1, na.rm=T),
              gender2 = sum(male ==0,na.rm=T)) %>%
  select(gender, gender2)
prop.table(gender)
tidy(chisq.test(gender))

## Ratings
ratings = birthorder %>%
  group_by(group_birthorder) %>%
  summarise(g_factor_mean = mean(g_factor_2015_old, na.rm=T),
            g_factor_confidence_low= mean(g_factor_2015_old, na.rm = T) - 
                (qt(.975, n()-1)*sd(g_factor_2015_old, na.rm = T)/sqrt(n())),
            g_factor_confidence_high = mean(g_factor_2015_old, na.rm = T) +
                (qt(.975, n()-1)*sd(g_factor_2015_old, na.rm = T)/sqrt(n())),
            big5_ext_mean = mean(big5_ext, na.rm=T),
            big5_ext_confidence_low= mean(big5_ext, na.rm = T) - 
                (qt(.975, n()-1)*sd(big5_ext, na.rm = T)/sqrt(n())),
            big5_ext_confidence_high = mean(big5_ext, na.rm = T) +
                (qt(.975, n()-1)*sd(big5_ext, na.rm = T)/sqrt(n())),
            big5_neu_mean = mean(big5_neu, na.rm=T),
            big5_neu_confidence_low= mean(big5_neu, na.rm = T) - 
                (qt(.975, n()-1)*sd(big5_neu, na.rm = T)/sqrt(n())),
            big5_neu_confidence_high = mean(big5_neu, na.rm = T) +
                (qt(.975, n()-1)*sd(big5_neu, na.rm = T)/sqrt(n())),
            big5_con_mean = mean(big5_con, na.rm=T),
            big5_con_confidence_low= mean(big5_con, na.rm = T) - 
                (qt(.975, n()-1)*sd(big5_con, na.rm = T)/sqrt(n())),
            big5_con_confidence_high = mean(big5_con, na.rm = T) +
                (qt(.975, n()-1)*sd(big5_con, na.rm = T)/sqrt(n())),
            big5_agree_mean = mean(big5_agree, na.rm=T),
            big5_agree_confidence_low= mean(big5_agree, na.rm = T) - 
                (qt(.975, n()-1)*sd(big5_agree, na.rm = T)/sqrt(n())),
            big5_agree_confidence_high = mean(big5_agree, na.rm = T) +
                (qt(.975, n()-1)*sd(big5_agree, na.rm = T)/sqrt(n())),
            big5_open_mean = mean(big5_open, na.rm=T),
            big5_open_confidence_low= mean(big5_open, na.rm = T) - 
                (qt(.975, n()-1)*sd(big5_open, na.rm = T)/sqrt(n())),
            big5_open_confidence_high = mean(big5_open, na.rm = T) +
                (qt(.975, n()-1)*sd(big5_open, na.rm = T)/sqrt(n())),
            riskA_mean = mean(riskA, na.rm=T),
            riskA_confidence_low= mean(riskA, na.rm = T) - 
                (qt(.975, n()-1)*sd(riskA, na.rm = T)/sqrt(n())),
            riskA_confidence_high = mean(riskA, na.rm = T) +
                (qt(.975, n()-1)*sd(riskA, na.rm = T)/sqrt(n())),
            riskB_mean = mean(riskB, na.rm=T),
            riskB_confidence_low= mean(riskB, na.rm = T) - 
                (qt(.975, n()-1)*sd(riskB, na.rm = T)/sqrt(n())),
            riskB_confidence_high = mean(riskB, na.rm = T) +
                (qt(.975, n()-1)*sd(riskB, na.rm = T)/sqrt(n())))
ratings

tidy(t.test(birthorder$g_factor_2015_old ~ birthorder$group_birthorder, var.equal = T))
cohen.d(birthorder$g_factor_2015_old, as.factor(birthorder$group_birthorder), na.rm = T)


tidy(t.test(birthorder$years_of_education ~ birthorder$group_birthorder, var.equal = T))
cohen.d(birthorder$years_of_education, as.factor(birthorder$group_birthorder), na.rm = T)


tidy(t.test(birthorder$big5_ext ~ birthorder$group_birthorder, var.equal = T))
cohen.d(birthorder$big5_ext, as.factor(birthorder$group_birthorder), na.rm = T)


tidy(t.test(birthorder$big5_neu ~ birthorder$group_birthorder, var.equal = T))
cohen.d(birthorder$big5_neu, as.factor(birthorder$group_birthorder), na.rm = T)


tidy(t.test(birthorder$big5_con ~ birthorder$group_birthorder, var.equal = T))
cohen.d(birthorder$big5_con, as.factor(birthorder$group_birthorder), na.rm = T)


tidy(t.test(birthorder$big5_agree ~ birthorder$group_birthorder, var.equal = T))
cohen.d(birthorder$big5_agree, as.factor(birthorder$group_birthorder), na.rm = T)


tidy(t.test(birthorder$big5_open ~ birthorder$group_birthorder, var.equal = T))
cohen.d(birthorder$big5_open, as.factor(birthorder$group_birthorder), na.rm = T)


tidy(t.test(birthorder$riskA ~ birthorder$group_birthorder, var.equal = T))
cohen.d(birthorder$riskA, as.factor(birthorder$group_birthorder), na.rm = T)

tidy(t.test(birthorder$riskB ~ birthorder$group_birthorder, var.equal = T))
cohen.d(birthorder$riskB, as.factor(birthorder$group_birthorder), na.rm = T)

## Educational Attainment
educational_attainment = birthorder %>%
  group_by(group_birthorder) %>%
  summarise(years_of_education_mean = mean(years_of_education, na.rm=T),
            years_of_education_confidence_low= mean(years_of_education, na.rm = T) - 
                (qt(.975, n()-1)*sd(years_of_education, na.rm = T)/sqrt(n())),
            years_of_education_high = mean(years_of_education, na.rm = T) +
                (qt(.975, n()-1)*sd(years_of_education, na.rm = T)/sqrt(n())))

educational_attainment

t.test(birthorder$years_of_education ~ birthorder$group_birthorder, var.equal = T)

ggplot(data=birthorder, aes(x=years_of_education, fill=group_birthorder)) +
  geom_histogram(stat="count", binwidth=.5, position="dodge")
```



## Descriptives
```{r correlation}
mean_sd = birthorder %>%
  group_by(group_birthorder) %>%
  summarise(age_mean = mean(age, na.rm=T),
            age_sd = sd(age, na.rm=T),
            g_factor_mean = mean(g_factor_2015_old, na.rm=T),
            g_factor_sd = sd(g_factor_2015_old, na.rm=T),
            big5_ext_mean = mean(big5_ext, na.rm=T),
            big5_ext_sd = sd(big5_ext, na.rm=T),
            big5_neu_mean = mean(big5_neu, na.rm=T),
            big5_neu_sd = sd(big5_neu, na.rm=T),
            big5_con_mean = mean(big5_con, na.rm=T),
            big5_con_sd = sd(big5_con, na.rm=T),
            big5_agree_mean = mean(big5_agree, na.rm=T),
            big5_agree_sd = sd(big5_agree, na.rm=T),
            big5_open_mean = mean(big5_open, na.rm=T),
            big5_open_sd = sd(big5_open, na.rm=T),
            riskA_mean = mean(riskA, na.rm=T),
            riskA_sd = sd(riskA, na.rm=T),
            riskB_mean = mean(riskB, na.rm=T),
            riskB_sd = sd(riskB, na.rm=T),
            years_of_education_mean = mean(years_of_education, na.rm=T),
            years_of_education_sd = sd(years_of_education, na.rm=T))
      
mean_sd
```

## Correlations
```{r}
cor = round(cor(birthorder %>% filter(group_birthorder == "test", group_outcome == "test") %>% ungroup() %>% select(age, male, g_factor_2015_old, years_of_education, big5_ext, big5_neu, big5_con, big5_agree, big5_open, riskA, riskB), use = "pairwise.complete.obs"), 2) 

x = birthorder %>% filter(group_birthorder == "test", group_outcome == "test")
cor.test(x$age, x$male, conf.level = 0.995)
```

## Reliability {.tabset}

### Intelligence
```{r}
alpha(birthorder %>% filter(group_birthorder == "test", group_outcome == "test") %>% ungroup() %>% select(raven_2015_old, math_2015_old, count_backwards, words_delayed, adaptive_numbering))
```

### Personality
```{r}
##Extraversion
alpha(birthorder %>% filter(group_birthorder == "test", group_outcome == "test") %>% ungroup() %>%
        select(e1, e2r_reversed, e3))

## Neuroticism
alpha(birthorder %>% filter(group_birthorder == "test", group_outcome == "test") %>% ungroup()
      %>% select(n1r_reversed, n2, n3))

##conscientiousness
alpha(birthorder %>% filter(group_birthorder == "test", group_outcome == "test") %>% ungroup() %>% select(c1, c2r_reversed, c3))

##Agreeableness
alpha(birthorder %>% filter(group_birthorder == "test", group_outcome == "test") %>% ungroup() %>% select(a1, a2, a3r_reversed))

##Openness
alpha(birthorder %>% filter(group_birthorder == "test", group_outcome == "test") %>% ungroup() %>% select(o1, o2, o3))
```


## Plots age and gender {.tabset}

### Plots Gender
```{r}
birthorder = birthorder %>% filter(age<=100)

birthorder = birthorder %>%
  mutate(outcome = age)
plot_gender(birthorder)

birthorder = birthorder %>%
  mutate(outcome = birthorder_genes)
plot_gender(birthorder)

birthorder = birthorder %>%
  mutate(outcome = sibling_count_genes)
plot_gender(birthorder)


birthorder = birthorder %>%
  mutate(outcome = g_factor_2015_old)
plot_gender(birthorder)

birthorder = birthorder %>%
  mutate(outcome = g_factor_2015_young)
plot_gender(birthorder)

birthorder = birthorder %>%
  mutate(outcome = g_factor_2007_old)
plot_gender(birthorder)

birthorder = birthorder %>%
  mutate(outcome = g_factor_2007_young)
plot_gender(birthorder)

birthorder = birthorder %>%
  mutate(outcome = big5_ext)
plot_gender(birthorder)

birthorder = birthorder %>%
  mutate(outcome = big5_con)
plot_gender(birthorder)

birthorder = birthorder %>%
  mutate(outcome = big5_open)
plot_gender(birthorder)

birthorder = birthorder %>%
  mutate(outcome = big5_neu)
plot_gender(birthorder)

birthorder = birthorder %>%
  mutate(outcome = big5_agree)
plot_gender(birthorder)

birthorder = birthorder %>%
  mutate(outcome = riskA)
plot_gender(birthorder)

birthorder = birthorder %>%
  mutate(outcome = riskB)
plot_gender(birthorder)

birthorder = birthorder %>%
  mutate(outcome = years_of_education)
plot_gender(birthorder)

plot_gender(birthorder %>% filter(!is.na(attended_school)) %>% mutate(outcome = as.numeric(attended_school)))

birthorder = birthorder %>%
  mutate(outcome = Elementary_missed)
plot_gender(birthorder)

birthorder = birthorder %>%
  mutate(outcome = Elementary_worked)
plot_gender(birthorder)

birthorder = birthorder %>%
  mutate(outcome = wage_last_month_log)
plot_gender(birthorder)

birthorder = birthorder %>%
  mutate(outcome = wage_last_year_log)
plot_gender(birthorder)

birthorder = birthorder %>%
  mutate(outcome = Self_employed)
plot_gender(birthorder)

birthorder = birthorder %>%
  mutate(outcome = ever_smoked)
plot_gender(birthorder)

```

### Plots Age
```{r}
plot_age(birthorder %>% filter(age<= 100) %>% mutate(outcome = birthorder_genes))


birthorder = birthorder %>%
  mutate(outcome = sibling_count_genes)
plot_age(birthorder)


birthorder = birthorder %>%
  mutate(outcome = g_factor_2015_old)
plot_age(birthorder)

birthorder = birthorder %>%
  mutate(outcome = g_factor_2015_young)
plot_age(birthorder)

birthorder = birthorder %>%
  mutate(outcome = g_factor_2007_old)
plot_age(birthorder)

birthorder = birthorder %>%
  mutate(outcome = g_factor_2007_young)
plot_age(birthorder)

birthorder = birthorder %>%
  mutate(outcome = big5_ext)
plot_age(birthorder)

birthorder = birthorder %>%
  mutate(outcome = big5_con)
plot_age(birthorder)

birthorder = birthorder %>%
  mutate(outcome = big5_open)
plot_age(birthorder)

birthorder = birthorder %>%
  mutate(outcome = big5_neu)
plot_age(birthorder)

birthorder = birthorder %>%
  mutate(outcome = big5_agree)
plot_age(birthorder)

birthorder = birthorder %>%
  mutate(outcome = riskA)
plot_age(birthorder)

birthorder = birthorder %>%
  mutate(outcome = riskB)
plot_age(birthorder)

birthorder = birthorder %>%
  mutate(outcome = years_of_education)
plot_age(birthorder)

plot_age(birthorder %>% filter(!is.na(attended_school)) %>% mutate(outcome = as.numeric(attended_school)))

birthorder = birthorder %>%
  mutate(outcome = Elementary_missed)
plot_age(birthorder)

birthorder = birthorder %>%
  mutate(outcome = Elementary_worked)
plot_age(birthorder)

birthorder = birthorder %>%
  mutate(outcome = wage_last_month_log)
plot_age(birthorder)

birthorder = birthorder %>%
  mutate(outcome = wage_last_year_log)
plot_age(birthorder)

birthorder = birthorder %>%
  mutate(outcome = Self_employed)
plot_age(birthorder)

birthorder = birthorder %>%
  mutate(outcome = ever_smoked)
plot_age(birthorder)

```
