==> 4_analyses_by_province.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
# <span style="color:#A6D854">Birth Order Effects</span> 

## Preparation {.tabset}

### Helper
```{r helper}
source("0_helpers.R")
opts_chunk$set(warning = FALSE)
```

### Load data
```{r Load Data}
birthorder = readRDS("data/alldata_birthorder.rds")
```

### Data preparations
```{r data preparations}
# For analyses we want to clean the dataset and get rid of all uninteresting data
birthorder = birthorder %>%
  filter(!is.na(pidlink)) %>% # no individuals who are only known from the pregnancy file
  filter(is.na(lifebirths) | lifebirths == 2) %>% # remove 7 and 2 individuals who are known as stillbirth or miscarriage but still have PID
  select(-lifebirths) %>%
  filter(!is.na(mother_pidlink)) %>%
  select(-father_pidlink) %>%
  filter(is.na(any_multiple_birth) | any_multiple_birth != 1) %>% # remove families with twins/triplets/..
  filter(!is.na(birthorder_naive)) %>%
  select(-starts_with("age_"), -wave, -any_multiple_birth, -multiple_birth) %>%
  mutate(money_spent_smoking_log = if_else(is.na(money_spent_smoking_log) & ever_smoked == 0, 0, money_spent_smoking_log),
         amount = if_else(is.na(amount) & ever_smoked == 0, 0, amount),
         amount_still_smokers = if_else(is.na(amount_still_smokers) &  still_smoking == 0, 0, amount_still_smokers),
         birthyear = lubridate::year(birthdate))

# recode Factor Variable as Dummy Variable
birthorder = left_join(birthorder,
                                birthorder %>%
                                  filter(!is.na(Category)) %>%
                                  mutate(var = 1) %>%
                                  select(pidlink, Category, var) %>%
                                  spread(Category, var, fill = 0, sep = "_"), by = "pidlink") %>%
  select(-Category)

# recode Factor Variable as Dummy Variable
birthorder = left_join(birthorder,
                                birthorder %>%
                                  filter(!is.na(Sector)) %>%
                                  mutate(var = 1) %>%
                                  select(pidlink, Sector, var) %>%
                                  spread(Sector, var, fill = 0, sep = "_"), by = "pidlink") %>%
  select(-Sector)

### Variables
birthorder = birthorder %>%
  mutate(
    # center variables that are used for analysis
  g_factor_2015_old = scale(g_factor_2015_old),
  g_factor_2015_young = scale(g_factor_2015_young),
  g_factor_2007_old = scale(g_factor_2007_old),
  g_factor_2007_young = scale(g_factor_2007_young),
  raven_2015_old = scale(raven_2015_old),
  math_2015_old = scale(math_2015_old),
  count_backwards = scale(count_backwards),
  raven_2015_young = scale(raven_2015_young),
  math_2015_young = scale(math_2015_young),
  words_remembered_avg = scale(words_remembered_avg),
  words_immediate = scale(words_immediate),
  words_delayed = scale(words_delayed),
  adaptive_numbering = scale(adaptive_numbering),
  raven_2007_old = scale(raven_2007_old),
  math_2007_old = scale(math_2007_old),
  raven_2007_young = scale(raven_2007_young),
  math_2007_young = scale(math_2007_young),
  riskA = scale(riskA),
  riskB = scale(riskB),
  years_of_education_z = scale(years_of_education),
  Total_score_highest_z = scale(Total_score_highest),
  wage_last_month_z = scale(wage_last_month_log),
  wage_last_year_z = scale(wage_last_year_log),
  big5_ext = scale(big5_ext),
  big5_con = scale(big5_con),
  big5_agree = scale(big5_agree),
  big5_open = scale(big5_open),
  big5_neu = scale(big5_neu),
  attended_school = as.integer(attended_school),
  attended_school = ifelse(attended_school == 1, 0,
                           ifelse(attended_school == 2, 1, NA)))

qplot(birthorder$male)
qplot(birthorder$g_factor_2015_old)
qplot(birthorder$g_factor_2015_old)

### Birthorder and Sibling Count
birthorder = birthorder %>% 
  mutate(
# birthorder as factors with levels of 1, 2, 3, 4, 5, 5+
    birthorder_naive_factor = as.character(birthorder_naive),
    birthorder_naive_factor = ifelse(birthorder_naive > 5, "5+",
                                            birthorder_naive_factor),
    birthorder_naive_factor = factor(birthorder_naive_factor, 
                                            levels = c("1","2","3","4","5","5+")),
    sibling_count_naive_factor = as.character(sibling_count_naive),
    sibling_count_naive_factor = ifelse(sibling_count_naive > 5, "5+",
                                               sibling_count_naive_factor),
    sibling_count_naive_factor = factor(sibling_count_naive_factor, 
                                               levels = c("2","3","4","5","5+")),

    birthorder_uterus_alive_factor = as.character(birthorder_uterus_alive),
    birthorder_uterus_alive_factor = ifelse(birthorder_uterus_alive > 5, "5+",
                                            birthorder_uterus_alive_factor),
    birthorder_uterus_alive_factor = factor(birthorder_uterus_alive_factor, 
                                            levels = c("1","2","3","4","5","5+")),
    sibling_count_uterus_alive_factor = as.character(sibling_count_uterus_alive),
    sibling_count_uterus_alive_factor = ifelse(sibling_count_uterus_alive > 5, "5+",
                                               sibling_count_uterus_alive_factor),
    sibling_count_uterus_alive_factor = factor(sibling_count_uterus_alive_factor, 
                                               levels = c("2","3","4","5","5+")),
    birthorder_uterus_preg_factor = as.character(birthorder_uterus_preg),
    birthorder_uterus_preg_factor = ifelse(birthorder_uterus_preg > 5, "5+",
                                           birthorder_uterus_preg_factor),
    birthorder_uterus_preg_factor = factor(birthorder_uterus_preg_factor,
                                           levels = c("1","2","3","4","5","5+")),
    sibling_count_uterus_preg_factor = as.character(sibling_count_uterus_preg),
    sibling_count_uterus_preg_factor = ifelse(sibling_count_uterus_preg > 5, "5+",
                                              sibling_count_uterus_preg_factor),
    sibling_count_uterus_preg_factor = factor(sibling_count_uterus_preg_factor, 
                                              levels = c("2","3","4","5","5+")),
    birthorder_genes_factor = as.character(birthorder_genes),
    birthorder_genes_factor = ifelse(birthorder_genes >5 , "5+", birthorder_genes_factor),
    birthorder_genes_factor = factor(birthorder_genes_factor, 
                                     levels = c("1","2","3","4","5","5+")),
    sibling_count_genes_factor = as.character(sibling_count_genes),
    sibling_count_genes_factor = ifelse(sibling_count_genes >5 , "5+",
                                        sibling_count_genes_factor),
    sibling_count_genes_factor = factor(sibling_count_genes_factor, 
                                        levels = c("2","3","4","5","5+")),
    # interaction birthorder * siblingcout for each birthorder
    count_birthorder_naive =
      factor(str_replace(as.character(interaction(birthorder_naive_factor,                                                              sibling_count_naive_factor)),
                        "\\.", "/"),
                                           levels =   c("1/2","2/2", "1/3",  "2/3",
                                                        "3/3", "1/4", "2/4", "3/4", "4/4",
                                                        "1/5", "2/5", "3/5", "4/5", "5/5",
                                                        "1/5+", "2/5+", "3/5+", "4/5+",
                                                        "5/5+", "5+/5+")),
    count_birthorder_uterus_alive =
      factor(str_replace(as.character(interaction(birthorder_uterus_alive_factor,                                                              sibling_count_uterus_alive_factor)),
                        "\\.", "/"),
                                           levels =   c("1/2","2/2", "1/3",  "2/3",
                                                        "3/3", "1/4", "2/4", "3/4", "4/4",
                                                        "1/5", "2/5", "3/5", "4/5", "5/5",
                                                        "1/5+", "2/5+", "3/5+", "4/5+",
                                                        "5/5+", "5+/5+")),
    count_birthorder_uterus_preg =
      factor(str_replace(as.character(interaction(birthorder_uterus_preg_factor,                                                              sibling_count_uterus_preg_factor)), 
                         "\\.", "/"),
                                           levels =   c("1/2","2/2", "1/3",  "2/3",
                                                        "3/3", "1/4", "2/4", "3/4", "4/4",
                                                        "1/5", "2/5", "3/5", "4/5", "5/5",
                                                        "1/5+", "2/5+", "3/5+", "4/5+",
                                                        "5/5+", "5+/5+")),
    count_birthorder_genes =
      factor(str_replace(as.character(interaction(birthorder_genes_factor,                                                              sibling_count_genes_factor)), "\\.", "/"),
                                           levels =   c("1/2","2/2", "1/3",  "2/3",
                                                        "3/3", "1/4", "2/4", "3/4", "4/4",
                                                        "1/5", "2/5", "3/5", "4/5", "5/5",
                                                        "1/5+", "2/5+", "3/5+", "4/5+",
                                                        "5/5+", "5+/5+")))

birthorder <- birthorder %>%
                     mutate(sibling_count = sibling_count_genes_factor,
                            birth_order_nonlinear = birthorder_genes_factor,
                            birth_order = birthorder_genes,
                            age = scale(age/10, scale = FALSE),
                            count_birth_order = count_birthorder_genes)
```

### By province {.active}

```{r}
birthorder <- birthorder %>% filter(!is.na(province))
birthorder$urban = factor(birthorder$sc05==1)
birthorder <- birthorder %>% drop_na(birth_order, mother_pidlink, age, male, urban) %>% group_by(province) %>% filter(n_nonmissing(years_of_education) > 50)

provinces <- birthorder %>% drop_na(birth_order, urban) %>%  select(province, urban, years_of_education, birth_order) %>% group_by(province) %>% 
  summarise(urban = mean(urban == TRUE), n_nonmissing(years_of_education), sd(birth_order, na.rm = TRUE)) %>% arrange(desc(urban))
kable(provinces)
```

```{r}
add_slope_by_province <- function(model) {
  m1 <- update(model, formula = . ~ . + (1  | province), REML = FALSE)
  m2 <- update(model, formula = . ~ . + (1 + birth_order || province), REML = FALSE)
  an <- anova(m1, m2)
  if (an$`Pr(>Chisq)`[2] < .01) {
    ranefs <- ranef(m2)$province %>% as.data.frame() %>% 
            tibble::rownames_to_column("province") %>% 
            left_join(provinces, by = "province") %>% 
            arrange(birth_order) %>% 
            mutate(birth_order = birth_order + fixef(m2)["birth_order"], 
                   province = forcats::fct_inorder(province)
                   ) 
    print(knitr::kable(ranefs %>% select(province, birth_order) %>% arrange(birth_order)))
    print(ranefs %>% 
            ggplot(aes(province, birth_order, colour = urban)) + 
            geom_point(size = 4) + 
            scale_colour_continuous(low = "red", high = "blue") +
            coord_flip() + 
            geom_hline(yintercept = fixef(m2)["birth_order"], linetype = 'dashed'))
    cat("\n\n\n")
  }
  an
}
```

Wherever adding a random slope for the linear birth order by province improved 
the model at a nominal significance of p < .05, we show how effects varied by 
province in a plot.


### Number of people by province
```{r}
qplot(forcats::fct_infreq(birthorder$province)) + coord_flip() + xlab("Province") + ylab("N")
```

### Family size by province

```{r}
ggplot(data = birthorder, aes(forcats::fct_infreq(province), sibling_count_genes)) + coord_flip() + xlab("Province") + ylab("N") + geom_pointrange(stat = 'summary')

qplot(sibling_count_genes, data = birthorder %>% group_by(province) %>% mutate(mean_sibling_count_genes= mean(sibling_count_genes, na.rm =T))) + geom_vline(aes(xintercept = mean_sibling_count_genes), color = 'red') + facet_wrap(~ forcats::fct_infreq(province), scales = "free_y") + coord_cartesian(xlim = c(0,10))
```


## Urbanisation by province
```{r}
ggplot(data = birthorder, aes(forcats::fct_infreq(province), if_else(urban == "TRUE", 1, 0))) + coord_flip() + xlab("Province") + ylab("N") + geom_pointrange(stat = 'summary')
```


### Urbanisation and family size
```{r}
ggplot(data = birthorder %>% group_by(province) %>% summarise(family_size = mean(sibling_count_genes, na.rm = TRUE), urban = mean(if_else(urban == "TRUE", 1,0), na.rm = T)), aes(family_size, urban, label = province)) + geom_text() + xlab("Family size") + ylab("Urbanisation")
```

## Intelligence {.active}

### g-factor 2015 old {}

```{r g-factor 2015 old}
birthorder <- birthorder %>% mutate(outcome = g_factor_2015_old)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count +
               (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```



### g-factor 2015 young {}

```{r g-factor 2015 young}
birthorder <- birthorder %>% mutate(outcome = g_factor_2015_young)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count +
               (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```


### g-factor 2007 old {}

```{r g-factor 2007 old}
birthorder <- birthorder %>% mutate(outcome = g_factor_2007_old)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

### g-factor 2007 young {}

```{r g-factor 2007 young}
birthorder <- birthorder %>% mutate(outcome = g_factor_2007_young)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```


### Raven 2015 old {}

```{r}
birthorder <- birthorder %>% mutate(outcome = raven_2015_old)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

### Numeracy 2015 old {}

```{r}
birthorder <- birthorder %>% mutate(outcome = math_2015_old)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

### Raven 2015 young {}

```{r}
birthorder <- birthorder %>% mutate(outcome = raven_2015_young)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

### Numeracy 2015 young {}

```{r}
birthorder <- birthorder %>% mutate(outcome = math_2015_young)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

### Raven 2007 old {}

```{r}
birthorder <- birthorder %>% mutate(outcome = raven_2007_old)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
             data = birthorder)
add_slope_by_province(model)
```

### Numeracy 2007 old{}

```{r}
birthorder <- birthorder %>% mutate(outcome = math_2007_old)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
             data = birthorder)
add_slope_by_province(model)
```

### Raven 2007 young {}

```{r}
birthorder <- birthorder %>% mutate(outcome = raven_2007_young)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
             data = birthorder)
add_slope_by_province(model)
```

### Numeracy 2007 young {}

```{r}
birthorder <- birthorder %>% mutate(outcome = math_2007_young)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
             data = birthorder)
add_slope_by_province(model)
```

### Counting backwards {}

```{r}
birthorder <- birthorder %>% mutate(outcome = count_backwards)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

### Immediate word recall {}

```{r}
birthorder <- birthorder %>% mutate(outcome = words_immediate)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

### Delayed word recall {}
```{r}
birthorder <- birthorder %>% mutate(outcome = words_delayed)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

### Adaptive Numbering {}

```{r}
birthorder <- birthorder %>% mutate(outcome = adaptive_numbering)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

## Personality {}

### Extraversion {}


```{r}
birthorder <- birthorder %>% mutate(outcome = big5_ext)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```


### Neuroticism {}


```{r}
birthorder <- birthorder %>% mutate(outcome = big5_neu)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```


### Conscientiousness {}


```{r}
birthorder <- birthorder %>% mutate(outcome = big5_con)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```


### Agreeableness {}


```{r}
birthorder <- birthorder %>% mutate(outcome = big5_agree)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```


### Openness {}


```{r}
birthorder <- birthorder %>% mutate(outcome = big5_open)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```


## Risk preference {}

### Risk A {}


```{r}
birthorder <- birthorder %>% mutate(outcome = riskA)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```


### Risk B {}


```{r}
birthorder <- birthorder %>% mutate(outcome = riskB)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```


## Educational Attainment {}
### Years of Education - z-standardized {}

```{r}
birthorder <- birthorder %>% mutate(outcome = years_of_education)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)

summary( m0_by_province <- lmer(outcome ~ sibling_count * urban + birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink) + (1 + birth_order | province), data = birthorder))
summary( m1_urbrur <- lmer(outcome ~ (sibling_count + birth_order) * urban + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink) + (1 + birth_order | province), data = birthorder))

library(brms)
summary( m0_by_province_ort <- brm(outcome ~ sibling_count * urban + birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink) + (1 + birth_order  + sibling_count|| province), data = birthorder, cores = 4, 
                               prior = c(prior(normal(0,2)), prior(normal(0,2), class = sd))))
summary( m0_by_province <- brm(outcome ~ sibling_count * urban + birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink) + (1 + birth_order  + sibling_count| province), data = birthorder, cores = 4, 
                               prior = c(prior(normal(0,2)), prior(normal(0,2), class = sd))))
summary( m1_urbrur <- brm(outcome ~ (sibling_count + birth_order) * urban  + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink) + (1 + birth_order  + sibling_count| province), data = birthorder, cores = 4, 
                               prior = c(prior(normal(0,2)), prior(normal(0,2), class = sd))))
loo(m0_by_province_ort, m0_by_province, m1_urbrur)

anova(m0_by_province, m1_urbrur)
plot(allEffects(m1_urbrur))
```

### Elementary missed {}

```{r}
birthorder <- birthorder %>% mutate(outcome = Elementary_missed)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```


### Elementary worked {}

```{r}
birthorder <- birthorder %>% mutate(outcome = Elementary_worked)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```


### Attended School {}

```{r}
birthorder <- birthorder %>% mutate(outcome = attended_school)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```


## Work {}

### Income Last Month (log) - z-standardized {}

```{r}
birthorder <- birthorder %>% mutate(outcome = wage_last_month_z)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

### Income last year (log) - z-standardized {}

```{r}
birthorder <- birthorder %>% mutate(outcome = wage_last_year_z)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

### Self-Employment - non standardized {}

```{r}
birthorder <- birthorder %>% mutate(outcome = Self_employed)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```


## Work Category{}

### Category_Casual worker in agriculture{}

```{r}
birthorder <- birthorder %>% mutate(outcome = `Category_Casual worker in agriculture`)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

### Category_Casual worker not in agriculture{}

```{r}
birthorder <- birthorder %>% mutate(outcome = `Category_Casual worker not in agriculture`)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

### Category_Government worker{}

```{r}
birthorder <- birthorder %>% mutate(outcome = `Category_Government worker`)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```


### Category_Private worker{}

```{r}
birthorder <- birthorder %>% mutate(outcome = `Category_Private worker`)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

### Category_Self-employed{}

```{r}
birthorder <- birthorder %>% mutate(outcome = `Category_Self-employed`)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

### Category_Unpaid family worker{}

```{r}
birthorder <- birthorder %>% mutate(outcome = `Category_Unpaid family worker`)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

## Work Sector {}

### Sector_Agriculture, forestry, fishing and hunting {}

```{r}
birthorder <- birthorder %>% mutate(outcome = `Sector_Agriculture, forestry, fishing and hunting`)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

### Sector_Construction {}

```{r}
birthorder <- birthorder %>% mutate(outcome = `Sector_Construction`)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

### Sector_Electricity, gas, water {}

```{r}
birthorder <- birthorder %>% mutate(outcome = `Sector_Electricity, gas, water`)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```


### Sector_Finance, insurance, real estate and business services {}

```{r}
birthorder <- birthorder %>% mutate(outcome = `Sector_Finance, insurance, real estate and business services`)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```


### Sector_Manufacturing {}

```{r}
birthorder <- birthorder %>% mutate(outcome = Sector_Manufacturing)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```



### Sector_Mining and quarrying {}

```{r}
birthorder <- birthorder %>% mutate(outcome = `Sector_Mining and quarrying`)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```


### Sector_Social services {}

```{r}
birthorder <- birthorder %>% mutate(outcome = `Sector_Social services`)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```


### Sector_Transportation, storage and communications {}

```{r}
birthorder <- birthorder %>% mutate(outcome = `Sector_Transportation, storage and communications`)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

## Smoking Behaviour {}
### ever_smoked {}

```{r}
birthorder <- birthorder %>% mutate(outcome = ever_smoked)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```

### still_smoking {}

```{r}
birthorder <- birthorder %>% mutate(outcome = still_smoking)
model = lmer(outcome ~ birth_order + poly(age, 3, raw = TRUE) + male + sibling_count + (1 | mother_pidlink),
                   data = birthorder)
add_slope_by_province(model)
```
