==> 4_analyses_plots.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---
# <span style="color:#A6D854">Birth Order Effects</span> {.tabset}

## Helper
```{r helper}
source("0_helpers.R")
opts_chunk$set(warning = FALSE)
```

## Load data
```{r Load Data}
birthorder = readRDS("data/alldata_birthorder.rds")
```

## Data preparations
```{r data preparations}
# For analyses we want to clean the dataset and get rid of all uninteresting data
birthorder = birthorder %>%
  filter(!is.na(pidlink)) %>% # no individuals who are only known from the pregnancy file
  filter(is.na(lifebirths) | lifebirths == 2) %>% # remove 7 and 2 individuals who are known as stillbirth or miscarriage but still have PID
  select(-lifebirths) %>%
  filter(!is.na(mother_pidlink)) %>%
  select(-father_pidlink) %>%
  filter(is.na(any_multiple_birth) | any_multiple_birth != 1) %>% # remove families with twins/triplets/..
  filter(!is.na(birthorder_naive)) %>%
  select(-starts_with("age_"), -wave, -any_multiple_birth, -multiple_birth) %>%
  mutate(money_spent_smoking_log = if_else(is.na(money_spent_smoking_log) & ever_smoked == 0, 0, money_spent_smoking_log),
         amount = if_else(is.na(amount) & ever_smoked == 0, 0, amount),
         amount_still_smokers = if_else(is.na(amount_still_smokers) &  still_smoking == 0, 0, amount_still_smokers),
         birthyear = lubridate::year(birthdate))

# recode Factor Variable as Dummy Variable
birthorder = left_join(birthorder,
                                birthorder %>%
                                  filter(!is.na(Category)) %>%
                                  mutate(var = 1) %>%
                                  select(pidlink, Category, var) %>%
                                  spread(Category, var, fill = 0, sep = "_"), by = "pidlink") %>%
  select(-Category)

# recode Factor Variable as Dummy Variable
birthorder = left_join(birthorder,
                                birthorder %>%
                                  filter(!is.na(Sector)) %>%
                                  mutate(var = 1) %>%
                                  select(pidlink, Sector, var) %>%
                                  spread(Sector, var, fill = 0, sep = "_"), by = "pidlink") %>%
  select(-Sector)


# data based on 4_analyses without running the analyses

ggplot(birthorder, aes(riskA)) + geom_bar() + xlab("Risk Aversion SetA" )
ggplot(birthorder, aes(riskB)) + geom_bar() + xlab("Risk Aversion SetB" )

cor.test(birthorder$riskA, birthorder$riskB)

ggplot(birthorder, aes(riskA, riskB)) + geom_jitter(alpha = .1) + xlab("Risk Aversion SetA" ) + ylab("Risk Aversion SetB" )

cor.test(birthorder$riskA, birthorder$ever_smoked)
cor.test(birthorder$riskB, birthorder$ever_smoked)

cor.test(birthorder$riskA, birthorder$still_smoking)
cor.test(birthorder$riskB, birthorder$still_smoking)


cor.test(birthorder$riskA, birthorder$Self_employed)
cor.test(birthorder$riskB, birthorder$Self_employed)

## Robust Analyses without weird categories
birthorder = birthorder %>%
  mutate(riskA_robust = ifelse(riskA == 5, NA, riskA),
         riskB_robust = ifelse(riskB == 1, NA, riskB))


ggplot(birthorder, aes(riskA_robust)) + geom_bar() + xlab("Risk Aversion SetA" )
ggplot(birthorder, aes(riskB_robust)) + geom_bar() + xlab("Risk Aversion SetB" )

cor.test(birthorder$riskA_robust, birthorder$riskB_robust)

ggplot(birthorder, aes(riskA_robust, riskB_robust)) + geom_jitter(alpha = .1) + xlab("Risk Aversion SetA" ) + ylab("Risk Aversion SetB" )

cor.test(birthorder$riskA_robust, birthorder$ever_smoked)
cor.test(birthorder$riskB_robust, birthorder$ever_smoked)

cor.test(birthorder$riskA_robust, birthorder$Self_employed)
cor.test(birthorder$riskB_robust, birthorder$Self_employed)

### Variables
birthorder = birthorder %>%
  mutate(
    # center variables that are used for analysis
  g_factor_2015_old = scale(g_factor_2015_old),
  g_factor_2015_young = scale(g_factor_2015_young),
  g_factor_2007_old = scale(g_factor_2007_old),
  g_factor_2007_young = scale(g_factor_2007_young),
  raven_2015_old = scale(raven_2015_old),
  math_2015_old = scale(math_2015_old),
  count_backwards = scale(count_backwards),
  raven_2015_young = scale(raven_2015_young),  years_of_education_z = scale(years_of_education),
  Total_score_highest_z = scale(Total_score_highest),

  math_2015_young = scale(math_2015_young),
  words_remembered_avg = scale(words_remembered_avg),
  words_immediate = scale(words_immediate),
  words_delayed = scale(words_delayed),
  adaptive_numbering = scale(adaptive_numbering),
  raven_2007_old = scale(raven_2007_old),
  math_2007_old = scale(math_2007_old),
  raven_2007_young = scale(raven_2007_young),
  math_2007_young = scale(math_2007_young),
  riskA = scale(riskA),
  riskB = scale(riskB),
  riskA_robust = scale(riskA_robust),
  riskB_robust = scale(riskB_robust),
  years_of_education_z = scale(years_of_education),
  Total_score_highest_z = scale(Total_score_highest),
  wage_last_month_z = scale(wage_last_month_log),
  wage_last_year_z = scale(wage_last_year_log),
  big5_ext = scale(big5_ext),
  big5_con = scale(big5_con),
  big5_agree = scale(big5_agree),
  big5_open = scale(big5_open),
  big5_neu = scale(big5_neu),
  attended_school = as.integer(attended_school),
  attended_school = ifelse(attended_school == 1, 0,
                           ifelse(attended_school == 2, 1, NA)))

qplot(birthorder$male)
qplot(birthorder$g_factor_2015_old)

### Birthorder and Sibling Count
birthorder = birthorder %>% 
  mutate(
# birthorder as factors with levels of 1, 2, 3, 4, 5, >5
    birthorder_naive_factor = as.character(birthorder_naive),
    birthorder_naive_factor = ifelse(birthorder_naive > 5, ">5",
                                            birthorder_naive_factor),
    birthorder_naive_factor = factor(birthorder_naive_factor, 
                                            levels = c("1","2","3","4","5",">5")),
    sibling_count_naive_factor = as.character(sibling_count_naive),
    sibling_count_naive_factor = ifelse(sibling_count_naive > 5, ">5",
                                               sibling_count_naive_factor),
    sibling_count_naive_factor = factor(sibling_count_naive_factor, 
                                               levels = c("2","3","4","5",">5")),

    birthorder_uterus_alive_factor = as.character(birthorder_uterus_alive),
    birthorder_uterus_alive_factor = ifelse(birthorder_uterus_alive > 5, ">5",
                                            birthorder_uterus_alive_factor),
    birthorder_uterus_alive_factor = factor(birthorder_uterus_alive_factor, 
                                            levels = c("1","2","3","4","5",">5")),
    sibling_count_uterus_alive_factor = as.character(sibling_count_uterus_alive),
    sibling_count_uterus_alive_factor = ifelse(sibling_count_uterus_alive > 5, ">5",
                                               sibling_count_uterus_alive_factor),
    sibling_count_uterus_alive_factor = factor(sibling_count_uterus_alive_factor, 
                                               levels = c("2","3","4","5",">5")),
    birthorder_uterus_preg_factor = as.character(birthorder_uterus_preg),
    birthorder_uterus_preg_factor = ifelse(birthorder_uterus_preg > 5, ">5",
                                           birthorder_uterus_preg_factor),
    birthorder_uterus_preg_factor = factor(birthorder_uterus_preg_factor,
                                           levels = c("1","2","3","4","5",">5")),
    sibling_count_uterus_preg_factor = as.character(sibling_count_uterus_preg),
    sibling_count_uterus_preg_factor = ifelse(sibling_count_uterus_preg > 5, ">5",
                                              sibling_count_uterus_preg_factor),
    sibling_count_uterus_preg_factor = factor(sibling_count_uterus_preg_factor, 
                                              levels = c("2","3","4","5",">5")),
    birthorder_genes_factor = as.character(birthorder_genes),
    birthorder_genes_factor = ifelse(birthorder_genes >5 , ">5", birthorder_genes_factor),
    birthorder_genes_factor = factor(birthorder_genes_factor, 
                                     levels = c("1","2","3","4","5",">5")),
    sibling_count_genes_factor = as.character(sibling_count_genes),
    sibling_count_genes_factor = ifelse(sibling_count_genes >5 , ">5",
                                        sibling_count_genes_factor),
    sibling_count_genes_factor = factor(sibling_count_genes_factor, 
                                        levels = c("2","3","4","5",">5")),
    # interaction birthorder * siblingcout for each birthorder
    count_birthorder_naive =
      factor(str_replace(as.character(interaction(birthorder_naive_factor,                                                              sibling_count_naive_factor)),
                        "\\.", "/"),
                                           levels =   c("1/2","2/2", "1/3",  "2/3",
                                                        "3/3", "1/4", "2/4", "3/4", "4/4",
                                                        "1/5", "2/5", "3/5", "4/5", "5/5",
                                                        "1/>5", "2/>5", "3/>5", "4/>5",
                                                        "5/>5", ">5/>5")),
    count_birthorder_uterus_alive =
      factor(str_replace(as.character(interaction(birthorder_uterus_alive_factor,                                                              sibling_count_uterus_alive_factor)),
                        "\\.", "/"),
                                           levels =   c("1/2","2/2", "1/3",  "2/3",
                                                        "3/3", "1/4", "2/4", "3/4", "4/4",
                                                        "1/5", "2/5", "3/5", "4/5", "5/5",
                                                        "1/>5", "2/>5", "3/>5", "4/>5",
                                                        "5/>5", ">5/>5")),
    count_birthorder_uterus_preg =
      factor(str_replace(as.character(interaction(birthorder_uterus_preg_factor,                                                              sibling_count_uterus_preg_factor)), 
                         "\\.", "/"),
                                           levels =   c("1/2","2/2", "1/3",  "2/3",
                                                        "3/3", "1/4", "2/4", "3/4", "4/4",
                                                        "1/5", "2/5", "3/5", "4/5", "5/5",
                                                        "1/>5", "2/>5", "3/>5", "4/>5",
                                                        "5/>5", ">5/>5")),
    count_birthorder_genes =
      factor(str_replace(as.character(interaction(birthorder_genes_factor,                                                              sibling_count_genes_factor)), "\\.", "/"),
                                           levels =   c("1/2","2/2", "1/3",  "2/3",
                                                        "3/3", "1/4", "2/4", "3/4", "4/4",
                                                        "1/5", "2/5", "3/5", "4/5", "5/5",
                                                        "1/>5", "2/>5", "3/>5", "4/>5",
                                                        "5/>5", ">5/>5")))

birthorder <- birthorder %>%
                     mutate(sibling_count = sibling_count_genes_factor,
                            birth_order_nonlinear = birthorder_genes_factor,
                            birth_order = birthorder_genes,
                            count_birth_order = count_birthorder_genes)

```




## Plots - nonadjusted y-axis
```{r}
plot_birthorder = function(model, ylabel = NULL, title = "", bo_var = "birth_order", separate = TRUE) {
  if(inherits(model, "merMod")) {
    varnames = names(model@frame)
  } else {
    varnames = names(model$model)
  }
  outcome = varnames[1]
  if(is.null(ylabel)) ylabel = outcome
  library(effects)
  library(tidyr)
  emm = allEffects(model)
  bo_var = names(emm)[names(emm) %contains% bo_var]
  cemm = as.data.frame(emm[[bo_var]])
  if (separate != TRUE) {
    cemm = cemm %>% rename_("Birth order" = bo_var) %>% mutate(Sibship = "across")
  } else {
    cemm = cemm %>%
      separate_(bo_var, into = c("Birth order", "Sibship"), sep = "/")
    number = spread(as.data.frame(table(model.frame(model)[`bo_var`])), Var1, Freq)
    n2 = paste0("2 (", sum(number$`1/2`, number$`2/2`), ")", seperate="")
    n3 = paste0("3 (", sum(number$`1/3`, number$`2/3`, number$`3/3`), ")",
                seperate="")
    n4 = paste0("4 (", sum(number$`1/4`, number$`2/4`, number$`3/4`,
                           number$`4/4`), ")", seperate="")
    n5 = paste0("5 (", sum(number$`1/5`, number$`2/5`, number$`3/5`,
                           number$`4/5`, number$`5/5`), ")", seperate="")
    n5more = paste0(">5 (", sum(number$`1/>5`, number$`2/>5`, number$`3/>5`,
                                number$`4/>5`, number$`5/>5`, number$`>5/>5`), ")",
                    seperate="")
    cemm = cemm %>%
      mutate(Sibship = recode_factor(Sibship, "2" = `n2`, "3" = `n3`, "4" = `n4`,
                                     "5" = `n5`, ">5" = `n5more`))
  }
  plotx = ggplot(cemm, aes(factor(`Birth order`,levels = c("1", "2", "3", "4", "5", ">5")),
                                  y = fit, ymax = upper, ymin = lower,
                           colour = `Sibship`, group = `Sibship`)) +
    geom_pointrange(stat = "identity", position = position_dodge(width = 0.5), size = 1) +
    geom_line(position = position_dodge(width = 0.5), size = 1) +
    scale_y_continuous(name = ylabel, limits = c(-0.4, 0.9),
                       breaks = c(-0.4, -0.2, 0, 0.2, 0.4, 0.6, 0.8)) +
    labs(title = title) +
    xlab("Birth order") +
    apatheme +
    theme(text = element_text(size=20), axis.text.x = element_text(size = 15),
          axis.text.y = element_text(size = 15), legend.text = element_text(size = 15),
          legend.title = element_text(size = 15),
          legend.background = element_rect(fill="gray90", size=.5, linetype="dotted"),
          plot.title = element_text(hjust = 0)) +
    guides(colour=guide_legend(title = "Sibship")) +
    scale_fill_brewer(palette = "Set2") +
    scale_colour_brewer(palette = "Set2")
  print(plotx)
  assign(paste0("plot_", outcome, seperate=""),plotx,.GlobalEnv)
}

library(grid)
library(gridExtra)

birthorder <- birthorder %>% mutate(outcome = g_factor_2015_old)
model = lmer(outcome ~ poly(age, 3, raw = TRUE) + male +
               count_birth_order +
               (1 | mother_pidlink),
                   data = birthorder)
ylabel = "Intelligence\n(z-standardized)"

plot_gfactor = plot_birthorder(model, ylabel = ylabel)  

birthorder <- birthorder %>% mutate(outcome = years_of_education_z)
model = lmer(outcome ~ poly(age, 3, raw = TRUE) + male +
               count_birth_order +
               (1 | mother_pidlink),
                   data = birthorder)
ylabel = "Years of education\n(z-standardized)"

plot_yoe = plot_birthorder(model, ylabel=ylabel) 

birthorder <- birthorder %>% mutate(outcome = big5_agree)
model = lmer(outcome ~ poly(age, 3, raw = TRUE) + male +
               count_birth_order +
               (1 | mother_pidlink),
                   data = birthorder)
ylabel = "Agreeableness\n(z-standardized)"

plot_big5_agree = plot_birthorder(model, ylabel=ylabel) 

birthorder <- birthorder %>% mutate(outcome = big5_ext)
model = lmer(outcome ~ poly(age, 3, raw = TRUE) + male +
               count_birth_order +
               (1 | mother_pidlink),
                   data = birthorder)
ylabel = "Extraversion\n(z-standardized)"

plot_big5_ext = plot_birthorder(model, ylabel=ylabel) 

birthorder <- birthorder %>% mutate(outcome = big5_con)
model = lmer(outcome ~ poly(age, 3, raw = TRUE) + male +
               count_birth_order +
               (1 | mother_pidlink),
                   data = birthorder)
ylabel = "Conscientiousness\n(z-standardized)"

plot_big5_con = plot_birthorder(model, ylabel=ylabel) 

birthorder <- birthorder %>% mutate(outcome = big5_open)
model = lmer(outcome ~ poly(age, 3, raw = TRUE) + male +
               count_birth_order +
               (1 | mother_pidlink),
                   data = birthorder)
ylabel = "Openness\n(z-standardized)"

plot_big5_open = plot_birthorder(model, ylabel=ylabel) 


birthorder <- birthorder %>% mutate(outcome = big5_neu)
model = lmer(outcome ~ poly(age, 3, raw = TRUE) + male +
               count_birth_order +
               (1 | mother_pidlink),
                   data = birthorder)
ylabel = "Neuroticism\n(z-standardized)"

plot_big5_neu = plot_birthorder(model, ylabel=ylabel) 

birthorder <- birthorder %>% mutate(outcome = riskA)
model = lmer(outcome ~ poly(age, 3, raw = TRUE) + male +
               count_birth_order +
               (1 | mother_pidlink),
                   data = birthorder)
ylabel = "Risk aversion A\n(z-standardized)"

plot_riskA = plot_birthorder(model, ylabel=ylabel) 

birthorder <- birthorder %>% mutate(outcome = riskB)
model = lmer(outcome ~ poly(age, 3, raw = TRUE) + male +
               count_birth_order +
               (1 | mother_pidlink),
                   data = birthorder)
ylabel = "Risk aversion B\n(z-standardized)"

plot_riskB = plot_birthorder(model, ylabel=ylabel) 


library(cowplot)
plots_nonadjusted = plot_grid(plot_gfactor, plot_yoe, plot_big5_ext, plot_big5_neu, plot_big5_con,
                     plot_big5_agree, plot_big5_open, plot_riskA, plot_riskB,
                     ncol = 3,
                     labels = c("(a)", "(b)", "(c)", "(d)", "(e)", "(f)", "(g)", "(h)", "(i)"))

pdf(file = "Submission EJP/Botzet_Figure2.pdf",   # The directory you want to save the file in
    width = 15, # The width of the plot in inches
    height = 17)
plots_nonadjusted
dev.off()

svg(file = "Submission EJP/Botzet_Figure2.svg",   # The directory you want to save the file in
    width = 15, # The width of the plot in inches
    height = 17)
plots_nonadjusted
dev.off() 

jpeg(file = "Submission EJP/Botzet_Figure2.jpg",   # The directory you want to save the file in
    width = 1500*0.8, # The width of the plot in inches
    height = 1700*0.8,
    quality = 200)
plots_nonadjusted
dev.off() 
```

## Plots -  adjusted y-axis
```{r}
plot_birthorder = function(model, ylabel = NULL, title = "", bo_var = "birth_order", separate = TRUE) {
  if(inherits(model, "merMod")) {
    varnames = names(model@frame)
  } else {
    varnames = names(model$model)
  }
  outcome = varnames[1]
  if(is.null(ylabel)) ylabel = outcome
  library(effects)
  library(tidyr)
  emm = allEffects(model)
  bo_var = names(emm)[names(emm) %contains% bo_var]
  cemm = as.data.frame(emm[[bo_var]])
  if (separate != TRUE) {
    cemm = cemm %>% rename_("Birth order" = bo_var) %>% mutate(Sibship = "across")
  } else {
    cemm = cemm %>%
      separate_(bo_var, into = c("Birth order", "Sibship"), sep = "/")
    number = spread(as.data.frame(table(model.frame(model)[`bo_var`])), Var1, Freq)
    n2 = paste0("2 (", sum(number$`1/2`, number$`2/2`), ")", seperate="")
    n3 = paste0("3 (", sum(number$`1/3`, number$`2/3`, number$`3/3`), ")",
                seperate="")
    n4 = paste0("4 (", sum(number$`1/4`, number$`2/4`, number$`3/4`,
                           number$`4/4`), ")", seperate="")
    n5 = paste0("5 (", sum(number$`1/5`, number$`2/5`, number$`3/5`,
                           number$`4/5`, number$`5/5`), ")", seperate="")
    n5more = paste0(">5 (", sum(number$`1/>5`, number$`2/>5`, number$`3/>5`,
                                number$`4/>5`, number$`5/>5`, number$`>5/>5`), ")",
                    seperate="")
    cemm = cemm %>%
      mutate(Sibship = recode_factor(Sibship, "2" = `n2`, "3" = `n3`, "4" = `n4`,
                                     "5" = `n5`, ">5" = `n5more`))
  }
  plotx = ggplot(cemm, aes(`Birth order`, y = fit, ymax = upper, ymin = lower,
                           colour = `Sibship`, group = `Sibship`)) +
    geom_pointrange(stat = "identity", position = position_dodge(width = 0.5), size = 1) +
    geom_line(position = position_dodge(width = 0.5), size = 1) +
    scale_y_continuous(name = ylabel) +
    labs(title= title) +
    apatheme +
    theme(text = element_text(size=20), axis.text.x = element_text(size = 15),
          axis.text.y = element_text(size = 15), legend.text = element_text(size = 15),
          legend.title = element_text(size = 15),
          legend.background = element_rect(fill="gray90", size=.5, linetype="dotted"),
          plot.title = element_text(hjust = 0)) +
    guides(colour=guide_legend(title = "Sibship")) +
    scale_fill_brewer(palette = "Set2") +
    scale_colour_brewer(palette = "Set2")
  print(plotx)
  assign(paste0("plot_", outcome, seperate=""),plotx,.GlobalEnv)
}

birthorder <- birthorder %>% mutate(outcome = g_factor_2015_old)
model = lmer(outcome ~ poly(age, 3, raw = TRUE) + male +
               count_birth_order +
               (1 | mother_pidlink),
                   data = birthorder)
ylabel = "Intelligence\n(z-standardized)"

plot_gfactor = plot_birthorder(model, ylabel=ylabel) 

birthorder <- birthorder %>% mutate(outcome = years_of_education_z)
model = lmer(outcome ~ poly(age, 3, raw = TRUE) + male +
               count_birth_order +
               (1 | mother_pidlink),
                   data = birthorder)
ylabel = "Years of education\n(z-standardized)"

plot_yoe = plot_birthorder(model, ylabel=ylabel) 

birthorder <- birthorder %>% mutate(outcome = big5_agree)
model = lmer(outcome ~ poly(age, 3, raw = TRUE) + male +
               count_birth_order +
               (1 | mother_pidlink),
                   data = birthorder)
ylabel = "Agreeableness\n(z-standardized)"

plot_big5_agree = plot_birthorder(model, ylabel=ylabel) 

birthorder <- birthorder %>% mutate(outcome = big5_ext)
model = lmer(outcome ~ poly(age, 3, raw = TRUE) + male +
               count_birth_order +
               (1 | mother_pidlink),
                   data = birthorder)
ylabel = "Extraversion\n(z-standardized)"

plot_big5_ext = plot_birthorder(model, ylabel=ylabel) 

birthorder <- birthorder %>% mutate(outcome = big5_con)
model = lmer(outcome ~ poly(age, 3, raw = TRUE) + male +
               count_birth_order +
               (1 | mother_pidlink),
                   data = birthorder)
ylabel = "Conscientiousness\n(z-standardized)"

plot_big5_con = plot_birthorder(model, ylabel=ylabel) 

birthorder <- birthorder %>% mutate(outcome = big5_open)
model = lmer(outcome ~ poly(age, 3, raw = TRUE) + male +
               count_birth_order +
               (1 | mother_pidlink),
                   data = birthorder)
ylabel = "Openness\n(z-standardized)"

plot_big5_open = plot_birthorder(model, ylabel=ylabel) 


birthorder <- birthorder %>% mutate(outcome = big5_neu)
model = lmer(outcome ~ poly(age, 3, raw = TRUE) + male +
               count_birth_order +
               (1 | mother_pidlink),
                   data = birthorder)
ylabel = "Neuroticism\n(z-standardized)"

plot_big5_neu = plot_birthorder(model, ylabel=ylabel) 

birthorder <- birthorder %>% mutate(outcome = riskA)
model = lmer(outcome ~ poly(age, 3, raw = TRUE) + male +
               count_birth_order +
               (1 | mother_pidlink),
                   data = birthorder)
ylabel = "Risk Aversion A\n(z-standardized)"

plot_riskA = plot_birthorder(model, ylabel=ylabel) 

birthorder <- birthorder %>% mutate(outcome = riskB)
model = lmer(outcome ~ poly(age, 3, raw = TRUE) + male +
               count_birth_order +
               (1 | mother_pidlink),
                   data = birthorder)
ylabel = "Risk Aversion B\n(z-standardized)"

plot_riskB = plot_birthorder(model, ylabel=ylabel) 

library(gridExtra)
plots_adjusted = grid.arrange(plot_gfactor, plot_yoe, plot_big5_ext, plot_big5_neu,
                              plot_big5_con,
                     plot_big5_agree, plot_big5_open, plot_riskA, plot_riskB,
                     ncol = 3)
```

## Effect Size Plots - linear


```{r g-factor 2015 old}
Intelligence = lmer(g_factor_2015_old ~ birthorder_genes + poly(age, 3, raw = TRUE) + male +
                        sibling_count_genes +
               (1 | mother_pidlink),
                   data = birthorder)

```


```{r g-factor 2015 old}
`Educational attainment` = lmer(years_of_education_z ~ birthorder_genes + poly(age, 3, raw = TRUE) + male +
                                  sibling_count_genes + (1 | mother_pidlink),
                   data = birthorder)


```



```{r g-factor 2015 old}
Extraversion = lmer(big5_ext ~ birthorder_genes + poly(age, 3, raw = TRUE) + male +
                                  sibling_count_genes + (1 | mother_pidlink),
                   data = birthorder)


```



```{r g-factor 2015 old}
Neuroticism = lmer(big5_neu ~ birthorder_genes + poly(age, 3, raw = TRUE) + male +
                                  sibling_count_genes + (1 | mother_pidlink),
                   data = birthorder)


```


```{r g-factor 2015 old}
Conscientiousness = lmer(big5_con ~ birthorder_genes + poly(age, 3, raw = TRUE) + male +
                                  sibling_count_genes + (1 | mother_pidlink),
                   data = birthorder)


```



```{r g-factor 2015 old}
Agreeableness = lmer(big5_agree ~ birthorder_genes + poly(age, 3, raw = TRUE) + male +
                                  sibling_count_genes + (1 | mother_pidlink),
                   data = birthorder)


```


```{r g-factor 2015 old}
Openness = lmer(big5_open ~ birthorder_genes + poly(age, 3, raw = TRUE) + male +
                                  sibling_count_genes + (1 | mother_pidlink),
                   data = birthorder)


```


```{r g-factor 2015 old}
`Risk A` = lmer(riskA ~ birthorder_genes + poly(age, 3, raw = TRUE) + male +
                                  sibling_count_genes + (1 | mother_pidlink),
                   data = birthorder)


```

```{r g-factor 2015 old}
`Risk B` = lmer(riskB ~ birthorder_genes + poly(age, 3, raw = TRUE) + male +
                                  sibling_count_genes + (1 | mother_pidlink),
                   data = birthorder)


```


## Multiplot - linear
```{r}
x = multiplot(Intelligence, `Educational attainment`, Agreeableness, Conscientiousness, Extraversion, Neuroticism, Openness, `Risk A`, `Risk B`, intercept = FALSE, coefficients = c("birthorder_genes", "sibling_count_genes"), plot = F)
        
x = data.frame(x)

x$Model = ifelse(x$Model == "`Educational attainment`", "Educational attainment", x$Model)
x$Model = ifelse(x$Model == "`Risk A`", "Risk aversion A", x$Model)
x$Model = ifelse(x$Model == "`Risk B`", "Risk aversion B", x$Model)

x$Model = as.factor(x$Model)
levels(x$Model)

x$Coefficient = ifelse(x$Coefficient == "birthorder_genes", "Birth order", "Sibship size")
x$Coefficients = x$Coefficient

plot = ggplot(x, aes(Model, Value)) +
  geom_pointrange(data=x, mapping=aes(x=Model, y=Value, ymin=HighOuter, ymax=LowOuter, color = Coefficients),
                  position = position_dodge(width=0.3), size = 1) +
  scale_colour_grey() +
  geom_hline(yintercept = 0, size = 1.5)  +
  geom_hline(yintercept = -.1, linetype = "dotted", size = 1.5) + 
  scale_x_discrete(limits=c("Risk aversion B", "Risk aversion A", "Openness", "Neuroticism", "Extraversion",
                            "Conscientiousness", "Agreeableness", "Educational attainment",
                            "Intelligence")) +
  scale_y_continuous(limits = c(-0.1, 0.05), breaks = c(-0.1, -0.05, 0, 0.05)) +
  coord_flip() +
  apatheme +
  theme(text = element_text(size=15), axis.text.x = element_text(size = 10),
          axis.text.y = element_text(size = 10), legend.text = element_text(size = 10),
          legend.title = element_text(size = 10),
          legend.background = element_rect(fill="gray90", size=.5, linetype="dotted"),
          plot.title = element_text(hjust = 0), legend.key.size = unit(2, 'line')) +
  labs(y = "Linear effect size", x = "Outcome")


pdf(file = "Submission EJP/Botzet_Figure1.pdf",   # The directory you want to save the file in
    width = 15, # The width of the plot in inches
    height = 17)
plot
dev.off()

svg(file = "Submission EJP/Botzet_Figure1.svg",   # The directory you want to save the file in
    width = 15, # The width of the plot in inches
    height = 17)
plots_nonadjusted
dev.off() 

jpeg(file = "Submission EJP/Botzet_Figure1.jpg",   # The directory you want to save the file in
    width = 1500*0.8, # The width of the plot in inches
    height = 1700*0.8,
    quality = 200)
plots_nonadjusted
dev.off() 
  
```


## Effect Size Plots - nonlinear
```{r g-factor 2015 old}
Intelligence = lmer(g_factor_2015_old ~ birthorder_genes_factor + poly(age, 3, raw = TRUE) + male +
                      sibling_count_genes_factor +
                      (1 | mother_pidlink),
                    data = birthorder)

```


```{r g-factor 2015 old}
`Educational Attainment` = lmer(years_of_education_z ~ birthorder_genes_factor + poly(age, 3, raw = TRUE) + male +
                                  sibling_count_genes_factor + (1 | mother_pidlink),
                                data = birthorder)


```



```{r g-factor 2015 old}
Extraversion = lmer(big5_ext ~ birthorder_genes_factor + poly(age, 3, raw = TRUE) + male +
                      sibling_count_genes_factor + (1 | mother_pidlink),
                    data = birthorder)


```



```{r g-factor 2015 old}
Neuroticism = lmer(big5_neu ~ birthorder_genes_factor + poly(age, 3, raw = TRUE) + male +
                     sibling_count_genes_factor + (1 | mother_pidlink),
                   data = birthorder)


```


```{r g-factor 2015 old}
Conscientiousness = lmer(big5_con ~ birthorder_genes_factor + poly(age, 3, raw = TRUE) + male +
                           sibling_count_genes_factor + (1 | mother_pidlink),
                         data = birthorder)


```



```{r g-factor 2015 old}
Agreeableness = lmer(big5_agree ~ birthorder_genes_factor + poly(age, 3, raw = TRUE) + male +
                       sibling_count_genes_factor + (1 | mother_pidlink),
                     data = birthorder)


```


```{r g-factor 2015 old}
Openness = lmer(big5_open ~ birthorder_genes_factor + poly(age, 3, raw = TRUE) + male +
                  sibling_count_genes_factor + (1 | mother_pidlink),
                data = birthorder)


```


```{r g-factor 2015 old}
RiskA = lmer(riskA ~ birthorder_genes_factor + poly(age, 3, raw = TRUE) + male +
               sibling_count_genes_factor + (1 | mother_pidlink),
             data = birthorder)


```

```{r g-factor 2015 old}
RiskB = lmer(riskB ~ birthorder_genes_factor + poly(age, 3, raw = TRUE) + male +
               sibling_count_genes_factor + (1 | mother_pidlink),
             data = birthorder)


```


## Multiplot - nonlinear
```{r}
x = multiplot(Intelligence, `Educational Attainment`, Agreeableness, Conscientiousness, Extraversion, Neuroticism, Openness, RiskA, RiskB, intercept = FALSE, coefficients = c("birthorder_genes_factor2", "birthorder_genes_factor3", "birthorder_genes_factor4", "birthorder_genes_factor5", "birthorder_genes_factor>5"), plot = F)

x = data.frame(x)

x$Model = ifelse(x$Model == "`Educational Attainment`", "Educational Attainment", x$Model)
x$Model = as.factor(x$Model)
levels(x$Model)

x$Coefficient = ifelse(x$Coefficient == "birthorder_genes_factor2", "2",
                       ifelse(x$Coefficient == "birthorder_genes_factor3", "3",
                              ifelse(x$Coefficient == "birthorder_genes_factor4", "4",
                                     ifelse(x$Coefficient == "birthorder_genes_factor5", "5",
                                            ifelse(x$Coefficient == "birthorder_genes_factor>5", ">5", NA)))))

colnames(x)[2] = "Birth order"



ggplot(x, aes(Model, Value)) +
  geom_pointrange(data=x, mapping=aes(x=Model, y=Value, ymin=HighOuter, ymax=LowOuter,
                                      color = `Birth order`),
                  position = position_dodge(width=0.3), size = 1) +
  scale_colour_brewer(palette = "Set2") +
  geom_hline(yintercept = 0, size = 1.5)  +
  geom_hline(yintercept = -.1, linetype = "dotted", size = 1.5) + 
  scale_x_discrete(limits=c("RiskB", "RiskA", "Openness", "Neuroticism", "Extraversion",
                            "Conscientiousness", "Agreeableness", "Educational Attainment",
                            "Intelligence")) +
  scale_y_continuous(limits = c(-0.36, 0.36), breaks = c(-0.3, -0.2, -0.1, 0, 0.1, 0.2, 0.3)) +
  coord_flip() +
  theme(text = element_text(size=25), axis.text.x = element_text(size = 20),
        axis.text.y = element_text(size = 20)) +
  labs(y = "Categorical effect size", x = "Outcome")
```
